refactor(ff-encode): split encoder_inner.rs into focused submodules
