--- crates/ff-encode/src/video/encoder_inner.rs
@@ -1,3756 +1,0 @@
-//! Internal video encoder implementation.
-//!
-//! This module contains the internal implementation details of the video encoder,
-//! including FFmpeg context management and encoding operations.
-
-// Rust 2024: Allow unsafe operations in unsafe functions for FFmpeg C API
-#![allow(unsafe_op_in_unsafe_fn)]
-// FFmpeg C API frequently requires raw pointer casting
-#![allow(clippy::ptr_as_ptr)]
-#![allow(clippy::cast_possible_wrap)]
-
-use crate::{AudioCodec, EncodeError, VideoCodec};
-use ff_format::{AudioFrame, VideoFrame};
-use ff_sys::{
-    AV_TIME_BASE, AVChannelLayout, AVChapter, AVCodecContext, AVCodecID, AVCodecID_AV_CODEC_ID_AAC,
-    AVCodecID_AV_CODEC_ID_AC3, AVCodecID_AV_CODEC_ID_ALAC, AVCodecID_AV_CODEC_ID_AV1,
-    AVCodecID_AV_CODEC_ID_DNXHD, AVCodecID_AV_CODEC_ID_DTS, AVCodecID_AV_CODEC_ID_EAC3,
-    AVCodecID_AV_CODEC_ID_FLAC, AVCodecID_AV_CODEC_ID_H264, AVCodecID_AV_CODEC_ID_HEVC,
-    AVCodecID_AV_CODEC_ID_MJPEG, AVCodecID_AV_CODEC_ID_MP3, AVCodecID_AV_CODEC_ID_MPEG2VIDEO,
-    AVCodecID_AV_CODEC_ID_MPEG4, AVCodecID_AV_CODEC_ID_NONE, AVCodecID_AV_CODEC_ID_OPUS,
-    AVCodecID_AV_CODEC_ID_PCM_S16LE, AVCodecID_AV_CODEC_ID_PCM_S24LE, AVCodecID_AV_CODEC_ID_PNG,
-    AVCodecID_AV_CODEC_ID_PRORES, AVCodecID_AV_CODEC_ID_VORBIS, AVCodecID_AV_CODEC_ID_VP8,
-    AVCodecID_AV_CODEC_ID_VP9, AVFormatContext, AVFrame, AVMediaType_AVMEDIA_TYPE_SUBTITLE,
-    AVPacket, AVPacketSideDataType_AV_PKT_DATA_CONTENT_LIGHT_LEVEL,
-    AVPacketSideDataType_AV_PKT_DATA_MASTERING_DISPLAY_METADATA, AVPixelFormat,
-    AVPixelFormat_AV_PIX_FMT_YUV420P, SwrContext, SwsContext, av_frame_alloc, av_frame_free,
-    av_interleaved_write_frame, av_mallocz, av_packet_alloc, av_packet_free,
-    av_packet_new_side_data, av_packet_unref, av_write_trailer, avcodec,
-    avformat_alloc_output_context2, avformat_free_context, avformat_new_stream,
-    avformat_write_header, swresample, swscale,
-};
-use std::ffi::CString;
-use std::ptr;
-
-/// FFmpeg `AVContentLightMetadata` — matches the C struct layout.
-///
-/// Not exposed by bindgen, so defined here with field order/types matching
-/// the FFmpeg 7.x header `libavutil/mastering_display_metadata.h`.
-#[repr(C)]
-struct AvContentLightMetadata {
-    /// Maximum Content Light Level (nits). Corresponds to `MaxCLL` in C.
-    max_cll: u32,
-    /// Maximum Frame-Average Light Level (nits). Corresponds to `MaxFALL` in C.
-    max_fall: u32,
-}
-
-/// FFmpeg `AVMasteringDisplayMetadata` — matches the C struct layout.
-///
-/// Not exposed by bindgen, so defined here with the exact field names and
-/// types from the FFmpeg 7.x header `libavutil/mastering_display_metadata.h`.
-#[repr(C)]
-struct AvMasteringDisplayMetadata {
-    /// Chromaticity coordinates of the source primaries:
-    /// `display_primaries[R=0/G=1/B=2][x=0/y=1]`.
-    display_primaries: [[ff_sys::AVRational; 2]; 3],
-    /// White point chromaticity: `white_point[x=0/y=1]`.
-    white_point: [ff_sys::AVRational; 2],
-    /// Minimum display luminance (AVRational with denominator 10000).
-    min_luminance: ff_sys::AVRational,
-    /// Maximum display luminance (AVRational with denominator 10000).
-    max_luminance: ff_sys::AVRational,
-    /// Flag: 1 if `display_primaries` is set, 0 otherwise.
-    has_primaries: std::os::raw::c_int,
-    /// Flag: 1 if `min_luminance`/`max_luminance` are set, 0 otherwise.
-    has_luminance: std::os::raw::c_int,
-}
-
-/// Maximum number of planes in AVFrame data/linesize arrays.
-///
-/// This corresponds to FFmpeg's `AV_NUM_DATA_POINTERS` (typically 8).
-/// Most pixel formats use 1-3 planes (e.g., RGB uses 1, YUV420P uses 3),
-/// but this allows for future extensibility and compatibility with
-/// exotic formats that may require more planes.
-const MAX_PLANES: usize = 8;
-
-/// FFmpeg pass-1 encoding flag: collect two-pass statistics, discard encoded output.
-const AV_CODEC_FLAG_PASS1: i32 = 512; // 1 << 9
-
-/// FFmpeg pass-2 encoding flag: use two-pass statistics from pass 1.
-const AV_CODEC_FLAG_PASS2: i32 = 1024; // 1 << 10
-
-/// Buffered raw frame data for two-pass re-encoding.
-///
-/// Stores the already-converted YUV420P plane data from pass 1 so that
-/// the same frames can be re-encoded in pass 2 without re-reading from
-/// the caller.
-pub(super) struct TwoPassFrame {
-    /// YUV420P plane data (Y plane at index 0, U at 1, V at 2).
-    planes: Vec<Vec<u8>>,
-    /// Linesize (stride) for each plane.
-    strides: Vec<usize>,
-    /// Frame width in pixels.
-    width: u32,
-    /// Frame height in pixels.
-    height: u32,
-    /// Presentation timestamp used when encoding this frame.
-    pts: i64,
-}
-
-/// Internal encoder state with FFmpeg contexts.
-pub(super) struct VideoEncoderInner {
-    /// Output format context
-    pub(super) format_ctx: *mut AVFormatContext,
-
-    /// Video codec context
-    pub(super) video_codec_ctx: Option<*mut AVCodecContext>,
-
-    /// Audio codec context (for future use)
-    pub(super) audio_codec_ctx: Option<*mut AVCodecContext>,
-
-    /// Video stream index
-    pub(super) video_stream_index: i32,
-
-    /// Audio stream index
-    pub(super) audio_stream_index: i32,
-
-    /// Scaling context for pixel format conversion
-    pub(super) sws_ctx: Option<*mut SwsContext>,
-
-    /// Resampling context for audio format conversion
-    pub(super) swr_ctx: Option<*mut SwrContext>,
-
-    /// Frame counter
-    pub(super) frame_count: u64,
-
-    /// Audio sample counter
-    pub(super) audio_sample_count: u64,
-
-    /// Bytes written
-    pub(super) bytes_written: u64,
-
-    /// Actual video codec name being used
-    pub(super) actual_video_codec: String,
-
-    /// Actual audio codec name being used
-    pub(super) actual_audio_codec: String,
-
-    /// Last source frame width (for SwsContext reuse optimization)
-    pub(super) last_src_width: Option<u32>,
-
-    /// Last source frame height (for SwsContext reuse optimization)
-    pub(super) last_src_height: Option<u32>,
-
-    /// Last source frame format (for SwsContext reuse optimization)
-    pub(super) last_src_format: Option<AVPixelFormat>,
-
-    /// Whether two-pass encoding is active.
-    pub(super) two_pass: bool,
-
-    /// Pass-1 codec context (two-pass mode only; None in single-pass and after pass 1 completes).
-    pub(super) pass1_codec_ctx: Option<*mut AVCodecContext>,
-
-    /// Buffered YUV420P frame data for pass-2 re-encoding (two-pass mode only).
-    pub(super) buffered_frames: Vec<TwoPassFrame>,
-
-    /// Stored configuration for reconstructing the pass-2 codec context.
-    pub(super) two_pass_config: Option<VideoEncoderConfig>,
-
-    /// Owned `stats_in` C string that must outlive the pass-2 codec context.
-    ///
-    /// Nulled out in `cleanup()` before `avcodec_free_context` to prevent FFmpeg
-    /// from calling `av_free` on a Rust-allocated pointer.
-    pub(super) stats_in_cstr: Option<std::ffi::CString>,
-
-    /// Subtitle passthrough info: (source_path, source_stream_index, output_stream_index).
-    ///
-    /// Set by `init_subtitle_passthrough`; read by `write_subtitle_packets`.
-    /// `None` if no subtitle passthrough was requested.
-    pub(super) subtitle_passthrough: Option<(String, usize, i32)>,
-
-    /// HDR10 static metadata to embed in keyframe packets.
-    ///
-    /// `None` if no HDR metadata was requested.
-    pub(super) hdr10_metadata: Option<ff_format::Hdr10Metadata>,
-}
-
-/// VideoEncoder configuration (stored from builder).
-#[derive(Debug, Clone)]
-pub(super) struct VideoEncoderConfig {
-    pub(super) path: std::path::PathBuf,
-    pub(super) video_width: Option<u32>,
-    pub(super) video_height: Option<u32>,
-    pub(super) video_fps: Option<f64>,
-    pub(super) video_codec: VideoCodec,
-    pub(super) video_bitrate_mode: Option<crate::BitrateMode>,
-    pub(super) preset: String,
-    pub(super) hardware_encoder: crate::HardwareEncoder,
-    pub(super) audio_sample_rate: Option<u32>,
-    pub(super) audio_channels: Option<u32>,
-    pub(super) audio_codec: AudioCodec,
-    pub(super) audio_bitrate: Option<u64>,
-    pub(super) _progress_callback: bool,
-    pub(super) two_pass: bool,
-    pub(super) metadata: Vec<(String, String)>,
-    pub(super) chapters: Vec<ff_format::chapter::ChapterInfo>,
-    pub(super) subtitle_passthrough: Option<(String, usize)>,
-    pub(super) codec_options: Option<crate::video::codec_options::VideoCodecOptions>,
-    pub(super) pixel_format: Option<ff_format::PixelFormat>,
-    pub(super) hdr10_metadata: Option<ff_format::Hdr10Metadata>,
-    pub(super) color_space: Option<ff_format::ColorSpace>,
-    pub(super) color_transfer: Option<ff_format::ColorTransfer>,
-    pub(super) color_primaries: Option<ff_format::ColorPrimaries>,
-    /// Binary attachments: (raw data, MIME type, filename).
-    pub(super) attachments: Vec<(Vec<u8>, String, String)>,
-    pub(super) container: Option<crate::OutputContainer>,
-}
-impl VideoEncoderInner {
-    /// Call `av_dict_set` for each metadata entry before `avformat_write_header`.
-    ///
-    /// # Safety
-    /// `format_ctx` must be a valid non-null pointer to an allocated `AVFormatContext`.
-    /// Must be called before `avformat_write_header`.
-    unsafe fn apply_metadata(
-        format_ctx: *mut ff_sys::AVFormatContext,
-        metadata: &[(String, String)],
-    ) {
-        for (key, value) in metadata {
-            let Ok(c_key) = std::ffi::CString::new(key.as_str()) else {
-                log::warn!("metadata key contains null byte, skipping key={key}");
-                continue;
-            };
-            let Ok(c_value) = std::ffi::CString::new(value.as_str()) else {
-                log::warn!("metadata value contains null byte, skipping key={key}");
-                continue;
-            };
-            // SAFETY: format_ctx is valid and non-null. c_key/c_value are valid
-            // CStrings covering this call. av_dict_set copies both strings.
-            let ret = ff_sys::av_dict_set(
-                &mut (*format_ctx).metadata,
-                c_key.as_ptr(),
-                c_value.as_ptr(),
-                0,
-            );
-            if ret < 0 {
-                log::warn!(
-                    "av_dict_set failed for metadata entry, skipping \
-                     key={key} error={}",
-                    ff_sys::av_error_string(ret)
-                );
-            }
-        }
-    }
-
-    /// Apply `movflags` for fMP4 containers before `avformat_write_header`.
-    ///
-    /// When `container` is [`crate::OutputContainer::FMp4`], sets
-    /// `movflags=+frag_keyframe+empty_moov+default_base_moof` via `av_opt_set`
-    /// on the format context's `priv_data`. This enables CMAF-compatible
-    /// fragmented output required for HLS fMP4 segments and MPEG-DASH.
-    ///
-    /// # Safety
-    /// `format_ctx` must be a valid non-null pointer to an allocated `AVFormatContext`
-    /// whose `priv_data` is non-null. Must be called before `avformat_write_header`.
-    unsafe fn apply_movflags(
-        format_ctx: *mut ff_sys::AVFormatContext,
-        container: Option<crate::OutputContainer>,
-    ) {
-        if container.is_some_and(|c| c.is_fragmented()) {
-            // SAFETY: format_ctx and priv_data are non-null; string literals are
-            // static and NUL-terminated. av_opt_set does not retain the pointers.
-            let ret = ff_sys::av_opt_set(
-                (*format_ctx).priv_data,
-                c"movflags".as_ptr(),
-                c"+frag_keyframe+empty_moov+default_base_moof".as_ptr(),
-                0,
-            );
-            if ret < 0 {
-                log::warn!(
-                    "av_opt_set movflags failed for fMP4 container error={}",
-                    ff_sys::av_error_string(ret)
-                );
-            }
-        }
-    }
-
-    /// Allocate `AVChapter` entries on the format context before `avformat_write_header`.
-    ///
-    /// # Safety
-    /// `format_ctx` must be a valid non-null pointer to an allocated `AVFormatContext`.
-    /// Must be called before `avformat_write_header`.
-    unsafe fn apply_chapters(
-        format_ctx: *mut ff_sys::AVFormatContext,
-        chapters: &[ff_format::chapter::ChapterInfo],
-    ) {
-        if chapters.is_empty() {
-            return;
-        }
-        let n = chapters.len();
-        // SAFETY: allocating an array of n pointers for the chapters field.
-        let chapters_arr =
-            av_mallocz(std::mem::size_of::<*mut AVChapter>() * n) as *mut *mut AVChapter;
-        if chapters_arr.is_null() {
-            log::warn!("av_mallocz failed for chapters array, skipping chapters");
-            return;
-        }
-        (*format_ctx).chapters = chapters_arr;
-        (*format_ctx).nb_chapters = 0;
-
-        for (i, chapter) in chapters.iter().enumerate() {
-            // SAFETY: allocating a zeroed AVChapter struct.
-            let chap = av_mallocz(std::mem::size_of::<AVChapter>()) as *mut AVChapter;
-            if chap.is_null() {
-                log::warn!(
-                    "av_mallocz failed for AVChapter, skipping chapter id={}",
-                    chapter.id()
-                );
-                continue;
-            }
-            // SAFETY: chap is freshly allocated, non-null, and zeroed.
-            (*chap).id = chapter.id();
-            (*chap).time_base = ff_sys::AVRational {
-                num: 1,
-                den: AV_TIME_BASE as i32,
-            };
-            (*chap).start = chapter.start().as_micros() as i64;
-            (*chap).end = chapter.end().as_micros() as i64;
-            (*chap).metadata = std::ptr::null_mut();
-
-            if let Some(title) = chapter.title() {
-                let Ok(c_title) = std::ffi::CString::new(title) else {
-                    log::warn!(
-                        "chapter title contains null byte, skipping title id={}",
-                        chapter.id()
-                    );
-                    // SAFETY: chapters_arr is valid with capacity n.
-                    *chapters_arr.add(i) = chap;
-                    (*format_ctx).nb_chapters += 1;
-                    continue;
-                };
-                // SAFETY: chap->metadata is null; av_dict_set allocates and copies.
-                let ret = ff_sys::av_dict_set(
-                    &mut (*chap).metadata,
-                    b"title\0".as_ptr() as *const _,
-                    c_title.as_ptr(),
-                    0,
-                );
-                if ret < 0 {
-                    log::warn!(
-                        "av_dict_set failed for chapter title, skipping title \
-                         id={} error={}",
-                        chapter.id(),
-                        ff_sys::av_error_string(ret)
-                    );
-                }
-            }
-            // SAFETY: i < n so the write is in bounds.
-            *chapters_arr.add(i) = chap;
-            (*format_ctx).nb_chapters += 1;
-        }
-    }
-
-    /// Apply per-codec options to an allocated (not yet opened) codec context.
-    ///
-    /// All `av_opt_set` return values are checked; a negative value is logged
-    /// as a warning and skipped — it never propagates as an error.
-    ///
-    /// # Safety
-    ///
-    /// `codec_ctx` must be a valid non-null pointer to an allocated
-    /// `AVCodecContext` whose `priv_data` has been set by
-    /// `avcodec_alloc_context3`. Must be called **before** `avcodec_open2`.
-    unsafe fn apply_codec_options(
-        codec_ctx: *mut AVCodecContext,
-        opts: &crate::video::codec_options::VideoCodecOptions,
-        encoder_name: &str,
-    ) {
-        use crate::video::codec_options::VideoCodecOptions;
-        use std::ffi::CString;
-
-        match opts {
-            VideoCodecOptions::H264(h264) => {
-                // profile
-                if let Ok(s) = CString::new(h264.profile.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null (set by avcodec_alloc_context3);
-                    // option name and value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"profile\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=profile value={} encoder={encoder_name}",
-                            h264.profile.as_str()
-                        );
-                    }
-                }
-                // level (only when explicitly set)
-                if let Some(level) = h264.level {
-                    let level_str = level.to_string();
-                    if let Ok(s) = CString::new(level_str.as_str()) {
-                        // SAFETY: codec_ctx and priv_data are non-null; option name and
-                        // value are valid NUL-terminated C strings.
-                        let ret = ff_sys::av_opt_set(
-                            (*codec_ctx).priv_data,
-                            b"level\0".as_ptr() as *const i8,
-                            s.as_ptr(),
-                            0,
-                        );
-                        if ret < 0 {
-                            log::warn!(
-                                "av_opt_set failed option=level value={level} \
-                                 encoder={encoder_name}"
-                            );
-                        }
-                    }
-                }
-                // Direct codec context fields.
-                // SAFETY: codec_ctx is a valid allocated AVCodecContext.
-                (*codec_ctx).max_b_frames = h264.bframes as i32;
-                (*codec_ctx).gop_size = h264.gop_size as i32;
-                (*codec_ctx).refs = h264.refs as i32;
-                // preset (libx264-specific; hardware encoders return a negative value
-                // which we log and skip — never returned as an error)
-                if let Some(preset) = h264.preset
-                    && let Ok(s) = CString::new(preset.as_str())
-                {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name
-                    // and value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"preset\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=preset value={} \
-                             encoder={encoder_name}",
-                            preset.as_str()
-                        );
-                    }
-                }
-                // tune (libx264-specific; hardware encoders return a negative value
-                // which we log and skip — never returned as an error)
-                if let Some(tune) = h264.tune
-                    && let Ok(s) = CString::new(tune.as_str())
-                {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name
-                    // and value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"tune\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=tune value={} \
-                             encoder={encoder_name}",
-                            tune.as_str()
-                        );
-                    }
-                }
-            }
-            VideoCodecOptions::H265(h265) => {
-                // profile
-                if let Ok(s) = CString::new(h265.profile.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name and
-                    // value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"profile\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=profile value={} encoder={encoder_name}",
-                            h265.profile.as_str()
-                        );
-                    }
-                }
-                // Auto-select yuv420p10le for Main10 (may be overridden by an explicit
-                // pixel_format() call applied after apply_codec_options returns).
-                if h265.profile == crate::video::codec_options::H265Profile::Main10 {
-                    // SAFETY: codec_ctx is valid; direct field write is safe.
-                    (*codec_ctx).pix_fmt = ff_sys::AVPixelFormat_AV_PIX_FMT_YUV420P10LE;
-                }
-                // tier
-                if let Ok(s) = CString::new(h265.tier.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name and
-                    // value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"tier\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=tier value={} encoder={encoder_name}",
-                            h265.tier.as_str()
-                        );
-                    }
-                }
-                // level (only when explicitly set)
-                if let Some(level) = h265.level {
-                    let level_str = level.to_string();
-                    if let Ok(s) = CString::new(level_str.as_str()) {
-                        // SAFETY: codec_ctx and priv_data are non-null; option name and
-                        // value are valid NUL-terminated C strings.
-                        let ret = ff_sys::av_opt_set(
-                            (*codec_ctx).priv_data,
-                            b"level\0".as_ptr() as *const i8,
-                            s.as_ptr(),
-                            0,
-                        );
-                        if ret < 0 {
-                            log::warn!(
-                                "av_opt_set failed option=level value={level} \
-                                 encoder={encoder_name}"
-                            );
-                        }
-                    }
-                }
-                // preset (libx265-specific; hardware HEVC encoders return a negative value
-                // which we log and skip — never returned as an error)
-                if let Some(ref preset) = h265.preset
-                    && let Ok(s) = CString::new(preset.as_str())
-                {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name
-                    // and value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"preset\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=preset value={preset} \
-                             encoder={encoder_name}"
-                        );
-                    }
-                }
-                // x265-params raw passthrough (libx265 only)
-                if let Some(ref params) = h265.x265_params
-                    && let Ok(s) = CString::new(params.as_str())
-                {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name
-                    // and value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"x265-params\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=x265-params value={params} \
-                             encoder={encoder_name}"
-                        );
-                    }
-                }
-            }
-            VideoCodecOptions::Av1(av1) => {
-                // cpu-used
-                let cpu_used_str = av1.cpu_used.to_string();
-                if let Ok(s) = CString::new(cpu_used_str.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name and
-                    // value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"cpu-used\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=cpu-used value={} encoder={encoder_name}",
-                            av1.cpu_used
-                        );
-                    }
-                }
-                // tile-rows
-                let tile_rows_str = av1.tile_rows.to_string();
-                if let Ok(s) = CString::new(tile_rows_str.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name and
-                    // value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"tile-rows\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=tile-rows value={} encoder={encoder_name}",
-                            av1.tile_rows
-                        );
-                    }
-                }
-                // tile-columns
-                let tile_cols_str = av1.tile_cols.to_string();
-                if let Ok(s) = CString::new(tile_cols_str.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name and
-                    // value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"tile-columns\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=tile-columns value={} \
-                             encoder={encoder_name}",
-                            av1.tile_cols
-                        );
-                    }
-                }
-                // usage
-                if let Ok(s) = CString::new(av1.usage.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name and
-                    // value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"usage\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=usage value={} encoder={encoder_name}",
-                            av1.usage.as_str()
-                        );
-                    }
-                }
-            }
-            VideoCodecOptions::Av1Svt(svt) => {
-                // preset (0–13)
-                let preset_str = svt.preset.to_string();
-                if let Ok(s) = CString::new(preset_str.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name and
-                    // value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"preset\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=preset value={} \
-                             encoder={encoder_name}",
-                            svt.preset
-                        );
-                    }
-                }
-                // tile-rows
-                let tile_rows_str = svt.tile_rows.to_string();
-                if let Ok(s) = CString::new(tile_rows_str.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name and
-                    // value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"tile_rows\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=tile_rows value={} \
-                             encoder={encoder_name}",
-                            svt.tile_rows
-                        );
-                    }
-                }
-                // tile-columns
-                let tile_cols_str = svt.tile_cols.to_string();
-                if let Ok(s) = CString::new(tile_cols_str.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name and
-                    // value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"tile_columns\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=tile_columns value={} \
-                             encoder={encoder_name}",
-                            svt.tile_cols
-                        );
-                    }
-                }
-                // svtav1-params raw passthrough
-                if let Some(ref params) = svt.svtav1_params
-                    && let Ok(s) = CString::new(params.as_str())
-                {
-                    // SAFETY: codec_ctx and priv_data are non-null; option name and
-                    // value are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"svtav1-params\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=svtav1-params value={params} \
-                             encoder={encoder_name}"
-                        );
-                    }
-                }
-            }
-            VideoCodecOptions::Vp9(vp9) => {
-                // CQ mode: override bitrate to 0 and set crf
-                if let Some(cq) = vp9.cq_level {
-                    // SAFETY: codec_ctx is non-null; direct field write is safe.
-                    (*codec_ctx).bit_rate = 0;
-                    let cq_str = cq.to_string();
-                    if let Ok(s) = CString::new(cq_str.as_str()) {
-                        // SAFETY: codec_ctx and priv_data are non-null; strings are
-                        // NUL-terminated.
-                        let ret = ff_sys::av_opt_set(
-                            (*codec_ctx).priv_data,
-                            b"crf\0".as_ptr() as *const i8,
-                            s.as_ptr(),
-                            0,
-                        );
-                        if ret < 0 {
-                            log::warn!(
-                                "av_opt_set failed option=crf value={cq} \
-                                 encoder={encoder_name}"
-                            );
-                        }
-                    }
-                }
-                // cpu-used
-                let cpu_used_str = vp9.cpu_used.to_string();
-                if let Ok(s) = CString::new(cpu_used_str.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null; strings are
-                    // NUL-terminated.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"cpu-used\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=cpu-used value={} \
-                             encoder={encoder_name}",
-                            vp9.cpu_used
-                        );
-                    }
-                }
-                // tile-columns
-                let tile_cols_str = vp9.tile_columns.to_string();
-                if let Ok(s) = CString::new(tile_cols_str.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null; strings are
-                    // NUL-terminated.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"tile-columns\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=tile-columns value={} \
-                             encoder={encoder_name}",
-                            vp9.tile_columns
-                        );
-                    }
-                }
-                // tile-rows
-                let tile_rows_str = vp9.tile_rows.to_string();
-                if let Ok(s) = CString::new(tile_rows_str.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null; strings are
-                    // NUL-terminated.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"tile-rows\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=tile-rows value={} \
-                             encoder={encoder_name}",
-                            vp9.tile_rows
-                        );
-                    }
-                }
-                // row-mt
-                let row_mt_str = if vp9.row_mt { "1" } else { "0" };
-                if let Ok(s) = CString::new(row_mt_str) {
-                    // SAFETY: codec_ctx and priv_data are non-null; strings are
-                    // NUL-terminated.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"row-mt\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=row-mt value={row_mt_str} \
-                             encoder={encoder_name}"
-                        );
-                    }
-                }
-            }
-            VideoCodecOptions::ProRes(prores) => {
-                // Set pixel format based on profile before avcodec_open2.
-                // 4444 profiles need yuva444p10le; 422 profiles need yuv422p10le.
-                // SAFETY: codec_ctx is non-null; direct field write is safe.
-                if prores.profile.is_4444() {
-                    (*codec_ctx).pix_fmt = ff_sys::AVPixelFormat_AV_PIX_FMT_YUVA444P10LE;
-                } else {
-                    (*codec_ctx).pix_fmt = ff_sys::AVPixelFormat_AV_PIX_FMT_YUV422P10LE;
-                }
-
-                // Apply profile via av_opt_set on priv_data.
-                let profile_str = prores.profile.profile_id().to_string();
-                if let Ok(s) = CString::new(profile_str.as_str()) {
-                    // SAFETY: codec_ctx and priv_data are non-null; strings are
-                    // NUL-terminated.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"profile\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=profile value={} \
-                             encoder={encoder_name}",
-                            prores.profile.profile_id()
-                        );
-                    }
-                }
-
-                // Apply optional vendor tag.
-                if let Some(vendor) = prores.vendor
-                    && let Ok(s) = CString::new(vendor.as_ref())
-                {
-                    // SAFETY: codec_ctx and priv_data are non-null; strings are
-                    // NUL-terminated.
-                    let ret = ff_sys::av_opt_set(
-                        (*codec_ctx).priv_data,
-                        b"vendor\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        0,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=vendor \
-                             encoder={encoder_name}"
-                        );
-                    }
-                }
-            }
-            VideoCodecOptions::Dnxhd(dnxhd) => {
-                use ff_format::PixelFormat;
-
-                // Set pixel format based on variant before avcodec_open2.
-                // SAFETY: codec_ctx is non-null; direct field write is safe.
-                (*codec_ctx).pix_fmt = match dnxhd.variant.pixel_format() {
-                    PixelFormat::Yuv422p => ff_sys::AVPixelFormat_AV_PIX_FMT_YUV422P,
-                    PixelFormat::Yuv422p10le => ff_sys::AVPixelFormat_AV_PIX_FMT_YUV422P10LE,
-                    PixelFormat::Yuv444p10le => ff_sys::AVPixelFormat_AV_PIX_FMT_YUV444P10LE,
-                    _ => ff_sys::AVPixelFormat_AV_PIX_FMT_YUV422P,
-                };
-
-                // For DNxHD variants, override bit_rate with the required fixed rate.
-                if let Some(bps) = dnxhd.variant.fixed_bitrate_bps() {
-                    // SAFETY: codec_ctx is non-null; direct field write is safe.
-                    (*codec_ctx).bit_rate = bps;
-                }
-
-                // Apply vprofile via av_opt_set on codec_ctx with AV_OPT_SEARCH_CHILDREN.
-                // Using the codec context (not priv_data) with SEARCH_CHILDREN allows the
-                // option to be found in child objects including priv_data.
-                if let Ok(s) = CString::new(dnxhd.variant.vprofile_str()) {
-                    // SAFETY: codec_ctx is non-null and cast to *mut c_void as required
-                    // by av_opt_set. AV_OPT_SEARCH_CHILDREN (1) searches child objects.
-                    // Strings are valid NUL-terminated C strings.
-                    let ret = ff_sys::av_opt_set(
-                        codec_ctx as *mut std::ffi::c_void,
-                        b"vprofile\0".as_ptr() as *const i8,
-                        s.as_ptr(),
-                        ff_sys::AV_OPT_SEARCH_CHILDREN as i32,
-                    );
-                    if ret < 0 {
-                        log::warn!(
-                            "av_opt_set failed option=vprofile value={} \
-                             encoder={encoder_name}",
-                            dnxhd.variant.vprofile_str()
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    /// Create a new encoder with the given configuration.
-    pub(super) fn new(config: &VideoEncoderConfig) -> Result<Self, EncodeError> {
-        unsafe {
-            ff_sys::ensure_initialized();
-
-            // Allocate output format context
-            let c_path = CString::new(config.path.to_str().ok_or_else(|| {
-                EncodeError::CannotCreateFile {
-                    path: config.path.clone(),
-                }
-            })?)
-            .map_err(|_| EncodeError::CannotCreateFile {
-                path: config.path.clone(),
-            })?;
-
-            // For image-sequence outputs (path contains '%'), use the `image2`
-            // muxer explicitly. The `image2` muxer manages I/O internally
-            // (AVFMT_NOFILE), so we must not call avio_open on it.
-            let is_image_sequence = config.path.to_str().is_some_and(|s| s.contains('%'));
-
-            let mut format_ctx: *mut AVFormatContext = ptr::null_mut();
-            let ret = avformat_alloc_output_context2(
-                &mut format_ctx,
-                ptr::null_mut(),
-                if is_image_sequence {
-                    b"image2\0".as_ptr() as *const i8
-                } else {
-                    ptr::null()
-                },
-                c_path.as_ptr(),
-            );
-
-            if ret < 0 || format_ctx.is_null() {
-                return Err(EncodeError::Ffmpeg {
-                    code: ret,
-                    message: format!(
-                        "Cannot create output context: {}",
-                        ff_sys::av_error_string(ret)
-                    ),
-                });
-            }
-
-            let mut encoder = Self {
-                format_ctx,
-                video_codec_ctx: None,
-                audio_codec_ctx: None,
-                video_stream_index: -1,
-                audio_stream_index: -1,
-                sws_ctx: None,
-                swr_ctx: None,
-                frame_count: 0,
-                audio_sample_count: 0,
-                bytes_written: 0,
-                actual_video_codec: String::new(),
-                actual_audio_codec: String::new(),
-                last_src_width: None,
-                last_src_height: None,
-                last_src_format: None,
-                two_pass: config.two_pass,
-                pass1_codec_ctx: None,
-                buffered_frames: Vec::new(),
-                two_pass_config: None,
-                stats_in_cstr: None,
-                subtitle_passthrough: None,
-                hdr10_metadata: config.hdr10_metadata.clone(),
-            };
-
-            // Initialize video encoder if configured
-            if let (Some(width), Some(height), Some(fps)) =
-                (config.video_width, config.video_height, config.video_fps)
-            {
-                encoder.init_video_encoder(
-                    width,
-                    height,
-                    fps,
-                    config.video_codec,
-                    config.video_bitrate_mode.as_ref(),
-                    &config.preset,
-                    config.hardware_encoder,
-                    config.two_pass,
-                    config.codec_options.as_ref(),
-                    config.pixel_format.as_ref(),
-                    config.color_space,
-                    config.color_transfer,
-                    config.color_primaries,
-                )?;
-            }
-
-            // Store config for pass-2 reconstruction (two-pass mode only).
-            if config.two_pass {
-                encoder.two_pass_config = Some(config.clone());
-            }
-
-            // Initialize audio encoder if configured
-            if let (Some(sample_rate), Some(channels)) =
-                (config.audio_sample_rate, config.audio_channels)
-            {
-                encoder.init_audio_encoder(
-                    sample_rate,
-                    channels,
-                    config.audio_codec,
-                    config.audio_bitrate,
-                )?;
-            }
-
-            // Register subtitle passthrough stream (must happen before avformat_write_header).
-            if let Some((ref path, stream_index)) = config.subtitle_passthrough {
-                encoder.init_subtitle_passthrough(path, stream_index);
-            }
-
-            // Register attachment streams (must happen before avformat_write_header).
-            if !config.attachments.is_empty() {
-                encoder.init_attachments(&config.attachments);
-            }
-
-            // For two-pass encoding the output file is opened in run_pass2() after
-            // pass-1 statistics have been collected.  Single-pass opens it now.
-            // Image-sequence output (path contains '%') uses the `image2` muxer which
-            // manages I/O internally (AVFMT_NOFILE) — skip avio_open in that case.
-            if !config.two_pass {
-                if !is_image_sequence {
-                    match ff_sys::avformat::open_output(
-                        &config.path,
-                        ff_sys::avformat::avio_flags::WRITE,
-                    ) {
-                        Ok(pb) => (*format_ctx).pb = pb,
-                        Err(_) => {
-                            encoder.cleanup();
-                            return Err(EncodeError::CannotCreateFile {
-                                path: config.path.clone(),
-                            });
-                        }
-                    }
-                }
-
-                Self::apply_movflags(format_ctx, config.container);
-                Self::apply_metadata(format_ctx, &config.metadata);
-                Self::apply_chapters(format_ctx, &config.chapters);
-                let ret = avformat_write_header(format_ctx, ptr::null_mut());
-                if ret < 0 {
-                    encoder.cleanup();
-                    return Err(EncodeError::Ffmpeg {
-                        code: ret,
-                        message: format!("Cannot write header: {}", ff_sys::av_error_string(ret)),
-                    });
-                }
-            }
-
-            Ok(encoder)
-        }
-    }
-
-    /// Initialize video encoder.
-    ///
-    /// When `two_pass` is `true` the codec context is opened with
-    /// `AV_CODEC_FLAG_PASS1` and stored in `pass1_codec_ctx`; in single-pass
-    /// mode it is stored in `video_codec_ctx` as usual.
-    unsafe fn init_video_encoder(
-        &mut self,
-        width: u32,
-        height: u32,
-        fps: f64,
-        codec: VideoCodec,
-        bitrate_mode: Option<&crate::BitrateMode>,
-        preset: &str,
-        hardware_encoder: crate::HardwareEncoder,
-        two_pass: bool,
-        codec_options: Option<&crate::video::codec_options::VideoCodecOptions>,
-        pixel_format: Option<&ff_format::PixelFormat>,
-        color_space: Option<ff_format::ColorSpace>,
-        color_transfer: Option<ff_format::ColorTransfer>,
-        color_primaries: Option<ff_format::ColorPrimaries>,
-    ) -> Result<(), EncodeError> {
-        use crate::BitrateMode;
-        // Select encoder based on codec and availability
-        let encoder_name = self.select_video_encoder(codec, hardware_encoder)?;
-        self.actual_video_codec = encoder_name.clone();
-
-        let c_encoder_name =
-            CString::new(encoder_name.as_str()).map_err(|_| EncodeError::Ffmpeg {
-                code: 0,
-                message: "Invalid encoder name".to_string(),
-            })?;
-
-        let codec_ptr =
-            avcodec::find_encoder_by_name(c_encoder_name.as_ptr()).ok_or_else(|| {
-                EncodeError::NoSuitableEncoder {
-                    codec: format!("{:?}", codec),
-                    tried: vec![encoder_name.clone()],
-                }
-            })?;
-
-        // Allocate codec context
-        let mut codec_ctx =
-            avcodec::alloc_context3(codec_ptr).map_err(EncodeError::from_ffmpeg_error)?;
-
-        // Configure codec context
-        (*codec_ctx).codec_id = codec_to_id(codec);
-        (*codec_ctx).width = width as i32;
-        (*codec_ctx).height = height as i32;
-        (*codec_ctx).time_base.num = 1;
-        (*codec_ctx).time_base.den = (fps * 1000.0) as i32; // Use millisecond precision
-        (*codec_ctx).framerate.num = fps as i32;
-        (*codec_ctx).framerate.den = 1;
-        (*codec_ctx).pix_fmt = AVPixelFormat_AV_PIX_FMT_YUV420P;
-
-        // Set bitrate control mode
-        match bitrate_mode {
-            Some(BitrateMode::Cbr(bps)) => {
-                (*codec_ctx).bit_rate = *bps as i64;
-            }
-            Some(BitrateMode::Vbr { target, max }) => {
-                (*codec_ctx).bit_rate = *target as i64;
-                (*codec_ctx).rc_max_rate = *max as i64;
-                (*codec_ctx).rc_buffer_size = (*max * 2) as i32;
-            }
-            Some(BitrateMode::Crf(q)) => {
-                let crf_str = CString::new(q.to_string()).map_err(|_| EncodeError::Ffmpeg {
-                    code: 0,
-                    message: "Invalid CRF value".to_string(),
-                })?;
-                // SAFETY: priv_data, option name, and value are all valid pointers
-                let ret = ff_sys::av_opt_set(
-                    (*codec_ctx).priv_data,
-                    b"crf\0".as_ptr() as *const i8,
-                    crf_str.as_ptr(),
-                    0,
-                );
-                if ret < 0 {
-                    log::warn!(
-                        "crf option not supported by encoder, falling back to default bitrate \
-                         encoder={encoder_name} crf={q}"
-                    );
-                    (*codec_ctx).bit_rate = 2_000_000;
-                }
-            }
-            None => {
-                // Default 2 Mbps
-                (*codec_ctx).bit_rate = 2_000_000;
-            }
-        }
-
-        // Set preset for x264/x265
-        if encoder_name.contains("264") || encoder_name.contains("265") {
-            let preset_cstr = CString::new(preset).map_err(|_| EncodeError::Ffmpeg {
-                code: 0,
-                message: "Invalid preset value".to_string(),
-            })?;
-            // SAFETY: priv_data, option name, and value are all valid pointers
-            let ret = ff_sys::av_opt_set(
-                (*codec_ctx).priv_data,
-                b"preset\0".as_ptr() as *const i8,
-                preset_cstr.as_ptr(),
-                0,
-            );
-            if ret < 0 {
-                log::warn!(
-                    "preset option not supported by encoder, ignoring \
-                     encoder={encoder_name} preset={preset}"
-                );
-            }
-        }
-
-        // Apply per-codec options before opening the codec context.
-        if let Some(opts) = codec_options {
-            // SAFETY: codec_ctx is valid and allocated; priv_data is set by
-            // avcodec_alloc_context3. Options are applied before avcodec_open2
-            // so they take effect during codec initialisation.
-            Self::apply_codec_options(codec_ctx, opts, &encoder_name);
-        }
-
-        // Apply explicit pixel format override (takes priority over codec-option auto-select).
-        if let Some(fmt) = pixel_format {
-            // SAFETY: codec_ctx is valid and allocated; direct field write is safe.
-            (*codec_ctx).pix_fmt = pixel_format_to_av(*fmt);
-        }
-
-        // Apply HDR10 color context: BT.2020 primaries, PQ transfer, BT.2020 NCL colorspace.
-        if self.hdr10_metadata.is_some() {
-            // SAFETY: codec_ctx is valid and allocated; direct field writes are safe.
-            (*codec_ctx).color_primaries = ff_sys::AVColorPrimaries_AVCOL_PRI_BT2020;
-            (*codec_ctx).color_trc = ff_sys::AVColorTransferCharacteristic_AVCOL_TRC_SMPTEST2084;
-            (*codec_ctx).colorspace = ff_sys::AVColorSpace_AVCOL_SPC_BT2020_NCL;
-        }
-
-        // Apply explicit color overrides (take priority over HDR10 automatic defaults).
-        if let Some(cs) = color_space {
-            // SAFETY: codec_ctx is valid and allocated; direct field write is safe.
-            (*codec_ctx).colorspace = color_space_to_av(cs);
-        }
-        if let Some(trc) = color_transfer {
-            // SAFETY: codec_ctx is valid and allocated; direct field write is safe.
-            (*codec_ctx).color_trc = color_transfer_to_av(trc);
-        }
-        if let Some(cp) = color_primaries {
-            // SAFETY: codec_ctx is valid and allocated; direct field write is safe.
-            (*codec_ctx).color_primaries = color_primaries_to_av(cp);
-        }
-
-        // For two-pass, set the pass-1 flag before opening the codec.
-        if two_pass {
-            // SAFETY: codec_ctx is a valid allocated (but not yet opened) context.
-            (*codec_ctx).flags |= AV_CODEC_FLAG_PASS1;
-        }
-
-        // Open codec
-        avcodec::open2(codec_ctx, codec_ptr, ptr::null_mut())
-            .map_err(EncodeError::from_ffmpeg_error)?;
-        let actual_pix_fmt = from_av_pixel_format((*codec_ctx).pix_fmt);
-        log::info!(
-            "codec opened codec={encoder_name} width={width} height={height} fps={fps} \
-             pix_fmt={actual_pix_fmt}"
-        );
-
-        // Create stream
-        let stream = avformat_new_stream(self.format_ctx, codec_ptr);
-        if stream.is_null() {
-            avcodec::free_context(&mut codec_ctx as *mut *mut _);
-            return Err(EncodeError::Ffmpeg {
-                code: 0,
-                message: "Cannot create stream".to_string(),
-            });
-        }
-
-        (*stream).time_base = (*codec_ctx).time_base;
-
-        // Copy codec parameters to stream
-        if !(*stream).codecpar.is_null() {
-            (*(*stream).codecpar).codec_id = (*codec_ctx).codec_id;
-            (*(*stream).codecpar).codec_type = ff_sys::AVMediaType_AVMEDIA_TYPE_VIDEO;
-            (*(*stream).codecpar).width = (*codec_ctx).width;
-            (*(*stream).codecpar).height = (*codec_ctx).height;
-            (*(*stream).codecpar).format = (*codec_ctx).pix_fmt;
-        }
-
-        self.video_stream_index = ((*self.format_ctx).nb_streams - 1) as i32;
-
-        // In two-pass mode the pass-1 context is stored separately; the real
-        // (pass-2) video_codec_ctx is initialised later in run_pass2().
-        if two_pass {
-            self.pass1_codec_ctx = Some(codec_ctx);
-        } else {
-            self.video_codec_ctx = Some(codec_ctx);
-        }
-
-        // Note: SwsContext initialization is deferred to convert_video_frame()
-        // for better optimization (skip unnecessary conversions, reuse context)
-
-        Ok(())
-    }
-
-    /// Select best available video encoder for the given codec.
-    ///
-    /// This method implements LGPL-compliant codec selection with automatic fallback:
-    /// - For H.264: Hardware encoders → libx264 (GPL only) → VP9 fallback
-    /// - For H.265: Hardware encoders → libx265 (GPL only) → AV1 fallback
-    /// - Hardware encoders (NVENC, QSV, AMF, VideoToolbox) are LGPL-compatible
-    /// - VP9 and AV1 are LGPL-compatible
-    fn select_video_encoder(
-        &self,
-        codec: VideoCodec,
-        hardware_encoder: crate::HardwareEncoder,
-    ) -> Result<String, EncodeError> {
-        // Early check: when Av1Svt is requested, verify that libsvtav1 is registered.
-        if codec == VideoCodec::Av1Svt {
-            // SAFETY: find_encoder_by_name is always safe to call with a valid NUL-terminated
-            // C string literal; the returned pointer is owned by FFmpeg and must not be freed.
-            let has_svt = unsafe {
-                avcodec::find_encoder_by_name(b"libsvtav1\0".as_ptr() as *const i8).is_some()
-            };
-            if !has_svt {
-                return Err(EncodeError::EncoderUnavailable {
-                    codec: "av1/svt".to_string(),
-                    hint: "Requires an FFmpeg build with --enable-libsvtav1 (LGPL)".to_string(),
-                });
-            }
-        }
-
-        // Early check: when H265 is requested, verify that at least one HEVC encoder
-        // is registered in this FFmpeg build before attempting candidate selection.
-        if codec == VideoCodec::H265 {
-            // SAFETY: avcodec::find_encoder is always safe to call with a valid codec ID;
-            // the returned pointer is owned by FFmpeg and must not be freed.
-            let has_hevc = unsafe { avcodec::find_encoder(AVCodecID_AV_CODEC_ID_HEVC).is_some() };
-            if !has_hevc {
-                return Err(EncodeError::EncoderUnavailable {
-                    codec: "h265/hevc".to_string(),
-                    hint: "Requires an FFmpeg build with HEVC encoder support \
-                           (hardware: hevc_nvenc/hevc_qsv/etc.; \
-                           software: --enable-libx265, GPL)"
-                        .to_string(),
-                });
-            }
-        }
-
-        let candidates: Vec<&str> = match codec {
-            VideoCodec::H264 => self.select_h264_encoder_candidates(hardware_encoder),
-            VideoCodec::H265 => self.select_h265_encoder_candidates(hardware_encoder),
-            VideoCodec::Vp9 => vec!["libvpx-vp9"],
-            VideoCodec::Av1 => vec!["libaom-av1", "libsvtav1", "av1"],
-            VideoCodec::Av1Svt => vec!["libsvtav1"],
-            VideoCodec::ProRes => vec!["prores_ks", "prores"],
-            VideoCodec::DnxHd => vec!["dnxhd"],
-            VideoCodec::Mpeg4 => vec!["mpeg4"],
-            VideoCodec::Vp8 => vec!["libvpx"],
-            VideoCodec::Mpeg2 => vec!["mpeg2video"],
-            VideoCodec::Mjpeg => vec!["mjpeg"],
-            VideoCodec::Png => vec!["png"],
-            _ => vec![],
-        };
-
-        // Try each candidate
-        for &name in &candidates {
-            unsafe {
-                let c_name = CString::new(name).map_err(|_| EncodeError::Ffmpeg {
-                    code: 0,
-                    message: "Invalid encoder name".to_string(),
-                })?;
-                if avcodec::find_encoder_by_name(c_name.as_ptr()).is_some() {
-                    return Ok(name.to_string());
-                }
-            }
-        }
-
-        Err(EncodeError::NoSuitableEncoder {
-            codec: format!("{:?}", codec),
-            tried: candidates.iter().map(|s| (*s).to_string()).collect(),
-        })
-    }
-
-    /// Select H.264 encoder candidates with LGPL compliance.
-    ///
-    /// Priority order:
-    /// 1. Hardware encoders (LGPL-compatible)
-    /// 2. libx264 (GPL only, requires `gpl` feature)
-    /// 3. VP9 fallback (LGPL-compatible)
-    fn select_h264_encoder_candidates(
-        &self,
-        hardware_encoder: crate::HardwareEncoder,
-    ) -> Vec<&'static str> {
-        let mut candidates = Vec::new();
-
-        // Add hardware encoders based on preference
-        #[cfg(feature = "hwaccel")]
-        match hardware_encoder {
-            crate::HardwareEncoder::Nvenc => {
-                candidates.extend_from_slice(&["h264_nvenc", "h264_qsv", "h264_amf"]);
-            }
-            crate::HardwareEncoder::Qsv => {
-                candidates.extend_from_slice(&["h264_qsv", "h264_nvenc", "h264_amf"]);
-            }
-            crate::HardwareEncoder::Amf => {
-                candidates.extend_from_slice(&["h264_amf", "h264_nvenc", "h264_qsv"]);
-            }
-            crate::HardwareEncoder::VideoToolbox => {
-                candidates.push("h264_videotoolbox");
-            }
-            crate::HardwareEncoder::Vaapi => {
-                candidates.push("h264_vaapi");
-            }
-            crate::HardwareEncoder::Auto => {
-                candidates.extend_from_slice(&[
-                    "h264_nvenc",
-                    "h264_qsv",
-                    "h264_amf",
-                    "h264_videotoolbox",
-                    "h264_vaapi",
-                ]);
-            }
-            crate::HardwareEncoder::None => {
-                // Skip hardware encoders
-            }
-        }
-
-        // Add GPL encoder if feature is enabled
-        #[cfg(feature = "gpl")]
-        {
-            candidates.push("libx264");
-        }
-
-        // Add LGPL-compatible fallback (VP9)
-        candidates.push("libvpx-vp9");
-
-        candidates
-    }
-
-    /// Select H.265 encoder candidates with LGPL compliance.
-    ///
-    /// Priority order:
-    /// 1. Hardware encoders (LGPL-compatible)
-    /// 2. libx265 (GPL only, requires `gpl` feature)
-    /// 3. AV1 fallback (LGPL-compatible)
-    fn select_h265_encoder_candidates(
-        &self,
-        hardware_encoder: crate::HardwareEncoder,
-    ) -> Vec<&'static str> {
-        let mut candidates = Vec::new();
-
-        // Add hardware encoders based on preference
-        #[cfg(feature = "hwaccel")]
-        match hardware_encoder {
-            crate::HardwareEncoder::Nvenc => {
-                candidates.extend_from_slice(&["hevc_nvenc", "hevc_qsv", "hevc_amf"]);
-            }
-            crate::HardwareEncoder::Qsv => {
-                candidates.extend_from_slice(&["hevc_qsv", "hevc_nvenc", "hevc_amf"]);
-            }
-            crate::HardwareEncoder::Amf => {
-                candidates.extend_from_slice(&["hevc_amf", "hevc_nvenc", "hevc_qsv"]);
-            }
-            crate::HardwareEncoder::VideoToolbox => {
-                candidates.push("hevc_videotoolbox");
-            }
-            crate::HardwareEncoder::Vaapi => {
-                candidates.push("hevc_vaapi");
-            }
-            crate::HardwareEncoder::Auto => {
-                candidates.extend_from_slice(&[
-                    "hevc_nvenc",
-                    "hevc_qsv",
-                    "hevc_amf",
-                    "hevc_videotoolbox",
-                    "hevc_vaapi",
-                ]);
-            }
-            crate::HardwareEncoder::None => {
-                // Skip hardware encoders
-            }
-        }
-
-        // Add GPL encoder if feature is enabled
-        #[cfg(feature = "gpl")]
-        {
-            candidates.push("libx265");
-        }
-
-        // Add LGPL-compatible fallback (AV1)
-        candidates.extend_from_slice(&["libaom-av1", "libsvtav1"]);
-
-        candidates
-    }
-
-    /// Initialize audio encoder.
-    unsafe fn init_audio_encoder(
-        &mut self,
-        sample_rate: u32,
-        channels: u32,
-        codec: AudioCodec,
-        bitrate: Option<u64>,
-    ) -> Result<(), EncodeError> {
-        // Select encoder based on codec and availability
-        let encoder_name = self.select_audio_encoder(codec)?;
-        self.actual_audio_codec = encoder_name.clone();
-
-        let c_encoder_name =
-            CString::new(encoder_name.as_str()).map_err(|_| EncodeError::Ffmpeg {
-                code: 0,
-                message: "Invalid encoder name".to_string(),
-            })?;
-
-        let codec_ptr =
-            avcodec::find_encoder_by_name(c_encoder_name.as_ptr()).ok_or_else(|| {
-                EncodeError::NoSuitableEncoder {
-                    codec: format!("{:?}", codec),
-                    tried: vec![encoder_name.clone()],
-                }
-            })?;
-
-        // Allocate codec context
-        let mut codec_ctx =
-            avcodec::alloc_context3(codec_ptr).map_err(EncodeError::from_ffmpeg_error)?;
-
-        // Configure codec context
-        (*codec_ctx).codec_id = audio_codec_to_id(codec);
-        (*codec_ctx).sample_rate = sample_rate as i32;
-
-        // Set channel layout using FFmpeg 7.x API
-        swresample::channel_layout::set_default(&mut (*codec_ctx).ch_layout, channels as i32);
-
-        // Set sample format (encoder's preferred format)
-        // We'll use FLTP (planar float) as it's widely supported
-        (*codec_ctx).sample_fmt = ff_sys::swresample::sample_format::FLTP;
-
-        // Set bitrate
-        if let Some(br) = bitrate {
-            (*codec_ctx).bit_rate = br as i64;
-        } else {
-            // Default bitrate based on codec
-            (*codec_ctx).bit_rate = match codec {
-                AudioCodec::Aac => 192_000,
-                AudioCodec::Opus => 128_000,
-                AudioCodec::Mp3 => 192_000,
-                AudioCodec::Flac => 0,  // Lossless
-                AudioCodec::Pcm => 0,   // Uncompressed
-                AudioCodec::Pcm16 => 0, // Uncompressed
-                AudioCodec::Pcm24 => 0, // Uncompressed
-                AudioCodec::Vorbis => 192_000,
-                AudioCodec::Ac3 => 192_000,
-                AudioCodec::Eac3 => 192_000,
-                AudioCodec::Dts => 0,  // Lossless/variable
-                AudioCodec::Alac => 0, // Lossless
-                _ => 192_000,
-            };
-        }
-
-        // Set time base
-        (*codec_ctx).time_base.num = 1;
-        (*codec_ctx).time_base.den = sample_rate as i32;
-
-        // Open codec
-        avcodec::open2(codec_ctx, codec_ptr, ptr::null_mut())
-            .map_err(EncodeError::from_ffmpeg_error)?;
-
-        // Create stream
-        let stream = avformat_new_stream(self.format_ctx, codec_ptr);
-        if stream.is_null() {
-            avcodec::free_context(&mut codec_ctx as *mut *mut _);
-            return Err(EncodeError::Ffmpeg {
-                code: 0,
-                message: "Cannot create stream".to_string(),
-            });
-        }
-
-        (*stream).time_base = (*codec_ctx).time_base;
-
-        // Copy codec parameters to stream
-        if !(*stream).codecpar.is_null() {
-            (*(*stream).codecpar).codec_id = (*codec_ctx).codec_id;
-            (*(*stream).codecpar).codec_type = ff_sys::AVMediaType_AVMEDIA_TYPE_AUDIO;
-            (*(*stream).codecpar).sample_rate = (*codec_ctx).sample_rate;
-            (*(*stream).codecpar).format = (*codec_ctx).sample_fmt;
-            // Copy channel layout
-            swresample::channel_layout::copy(
-                &mut (*(*stream).codecpar).ch_layout,
-                &(*codec_ctx).ch_layout,
-            )
-            .map_err(EncodeError::from_ffmpeg_error)?;
-        }
-
-        self.audio_stream_index = ((*self.format_ctx).nb_streams - 1) as i32;
-        self.audio_codec_ctx = Some(codec_ctx);
-
-        Ok(())
-    }
-
-    /// Select best available audio encoder for the given codec.
-    fn select_audio_encoder(&self, codec: AudioCodec) -> Result<String, EncodeError> {
-        let candidates: Vec<&str> = match codec {
-            AudioCodec::Aac => vec!["aac", "libfdk_aac"],
-            AudioCodec::Opus => vec!["libopus"],
-            AudioCodec::Mp3 => vec!["libmp3lame", "mp3"],
-            AudioCodec::Flac => vec!["flac"],
-            AudioCodec::Pcm => vec!["pcm_s16le"],
-            AudioCodec::Pcm16 => vec!["pcm_s16le"],
-            AudioCodec::Pcm24 => vec!["pcm_s24le"],
-            AudioCodec::Vorbis => vec!["libvorbis", "vorbis"],
-            AudioCodec::Ac3 => vec!["ac3"],
-            AudioCodec::Eac3 => vec!["eac3"],
-            AudioCodec::Dts => vec![],
-            AudioCodec::Alac => vec!["alac"],
-            _ => vec![],
-        };
-
-        // Try each candidate
-        for &name in &candidates {
-            unsafe {
-                let c_name = CString::new(name).map_err(|_| EncodeError::Ffmpeg {
-                    code: 0,
-                    message: "Invalid encoder name".to_string(),
-                })?;
-                if avcodec::find_encoder_by_name(c_name.as_ptr()).is_some() {
-                    return Ok(name.to_string());
-                }
-            }
-        }
-
-        Err(EncodeError::NoSuitableEncoder {
-            codec: format!("{:?}", codec),
-            tried: candidates.iter().map(|s| (*s).to_string()).collect(),
-        })
-    }
-
-    /// Push a video frame for encoding.
-    ///
-    /// In two-pass mode the frame is converted to YUV420P via the pass-1 codec
-    /// context, the converted data is buffered for pass-2 replay, and the frame
-    /// is then sent through the pass-1 encoder (whose output is discarded).
-    pub(super) fn push_video_frame(&mut self, frame: &VideoFrame) -> Result<(), EncodeError> {
-        // SAFETY: self is properly initialised; all raw FFmpeg pointers are valid and exclusively owned.
-        unsafe {
-            // ── Two-pass path ────────────────────────────────────────────────────
-            if self.two_pass {
-                let pass1_ctx = self
-                    .pass1_codec_ctx
-                    .ok_or_else(|| EncodeError::InvalidConfig {
-                        reason: "Pass-1 codec context not initialized".to_string(),
-                    })?;
-
-                // Convert the incoming frame to YUV420P (the pass-1 codec's format).
-                let mut av_frame = av_frame_alloc();
-                if av_frame.is_null() {
-                    return Err(EncodeError::Ffmpeg {
-                        code: 0,
-                        message: "Cannot allocate frame".to_string(),
-                    });
-                }
-
-                let convert_result = self.convert_video_frame(frame, av_frame, pass1_ctx);
-                if let Err(e) = convert_result {
-                    av_frame_free(&mut av_frame as *mut *mut _);
-                    return Err(e);
-                }
-
-                // Buffer the converted YUV420P data for pass-2 replay.
-                let width = (*pass1_ctx).width as u32;
-                let height = (*pass1_ctx).height as u32;
-                let uv_height = (height as usize).div_ceil(2);
-
-                let planes: Vec<Vec<u8>> = (0..3)
-                    .map(|i| {
-                        if (*av_frame).data[i].is_null() {
-                            return Vec::new();
-                        }
-                        let stride = (*av_frame).linesize[i] as usize;
-                        let h = if i == 0 { height as usize } else { uv_height };
-                        // SAFETY: data[i] points to a valid buffer of stride * h bytes
-                        // allocated by av_frame_get_buffer inside convert_video_frame.
-                        std::slice::from_raw_parts((*av_frame).data[i], stride * h).to_vec()
-                    })
-                    .collect();
-
-                let strides: Vec<usize> =
-                    (0..3).map(|i| (*av_frame).linesize[i] as usize).collect();
-
-                self.buffered_frames.push(TwoPassFrame {
-                    planes,
-                    strides,
-                    width,
-                    height,
-                    pts: self.frame_count as i64,
-                });
-
-                // Send to pass-1 encoder and discard the encoded output.
-                (*av_frame).pts = self.frame_count as i64;
-                let send_result = avcodec::send_frame(pass1_ctx, av_frame);
-                if let Err(e) = send_result {
-                    av_frame_free(&mut av_frame as *mut *mut _);
-                    return Err(EncodeError::Ffmpeg {
-                        code: e,
-                        message: format!(
-                            "Failed to send frame to pass-1 encoder: {}",
-                            ff_sys::av_error_string(e)
-                        ),
-                    });
-                }
-
-                let drain_result = self.drain_pass1_packets(pass1_ctx);
-                av_frame_free(&mut av_frame as *mut *mut _);
-                drain_result?;
-
-                self.frame_count += 1;
-                return Ok(());
-            }
-
-            // ── Single-pass path ─────────────────────────────────────────────────
-            let codec_ctx = self
-                .video_codec_ctx
-                .ok_or_else(|| EncodeError::InvalidConfig {
-                    reason: "Video codec not initialized".to_string(),
-                })?;
-
-            // Allocate AVFrame
-            let mut av_frame = av_frame_alloc();
-            if av_frame.is_null() {
-                return Err(EncodeError::Ffmpeg {
-                    code: 0,
-                    message: "Cannot allocate frame".to_string(),
-                });
-            }
-
-            // Convert VideoFrame to AVFrame
-            let convert_result = self.convert_video_frame(frame, av_frame, codec_ctx);
-            if let Err(e) = convert_result {
-                av_frame_free(&mut av_frame as *mut *mut _);
-                return Err(e);
-            }
-
-            // Set frame properties
-            (*av_frame).pts = self.frame_count as i64;
-
-            // Send frame to encoder
-            let send_result = avcodec::send_frame(codec_ctx, av_frame);
-            if let Err(e) = send_result {
-                av_frame_free(&mut av_frame as *mut *mut _);
-                return Err(EncodeError::Ffmpeg {
-                    code: e,
-                    message: format!("Failed to send frame: {}", ff_sys::av_error_string(e)),
-                });
-            }
-
-            // Receive packets
-            let receive_result = self.receive_packets();
-
-            // Always cleanup the frame
-            av_frame_free(&mut av_frame as *mut *mut _);
-
-            // Check if receiving packets failed
-            receive_result?;
-
-            self.frame_count += 1;
-
-            Ok(())
-        } // unsafe
-    }
-
-    /// Drain and discard all pending packets from a codec context.
-    ///
-    /// Used during pass-1 of two-pass encoding to prevent the packet queue
-    /// from filling up without writing any data to the output file.
-    ///
-    /// # Safety
-    ///
-    /// `codec_ctx` must be a valid, open `AVCodecContext`.
-    unsafe fn drain_pass1_packets(
-        &self,
-        codec_ctx: *mut AVCodecContext,
-    ) -> Result<(), EncodeError> {
-        let mut packet = av_packet_alloc();
-        if packet.is_null() {
-            return Err(EncodeError::Ffmpeg {
-                code: 0,
-                message: "Cannot allocate packet".to_string(),
-            });
-        }
-
-        loop {
-            match avcodec::receive_packet(codec_ctx, packet) {
-                Ok(()) => {
-                    // Discard — do not write to the format context.
-                    av_packet_unref(packet);
-                }
-                Err(e) if e == ff_sys::error_codes::EAGAIN || e == ff_sys::error_codes::EOF => {
-                    break;
-                }
-                Err(e) => {
-                    av_packet_free(&mut packet as *mut *mut _);
-                    return Err(EncodeError::Ffmpeg {
-                        code: e,
-                        message: format!(
-                            "Error receiving packet from pass-1 encoder: {}",
-                            ff_sys::av_error_string(e)
-                        ),
-                    });
-                }
-            }
-        }
-
-        av_packet_free(&mut packet as *mut *mut _);
-        Ok(())
-    }
-
-    /// Convert VideoFrame to AVFrame with pixel format conversion if needed.
-    ///
-    /// This method implements several optimizations in priority order:
-    /// 1. **Fast path**: Skips conversion entirely if format/dimensions match
-    /// 2. **Context reuse**: Reuses SwsContext when source properties unchanged
-    /// 3. **Lazy init**: Reinitializes SwsContext only when needed
-    /// 4. **Fast algorithm**: Uses BILINEAR scaling for speed/quality balance
-    ///
-    /// The caller supplies `codec_ctx` explicitly so this function can be used
-    /// with both the regular `video_codec_ctx` and the pass-1 `pass1_codec_ctx`.
-    ///
-    /// # Performance Characteristics
-    ///
-    /// - Same format/size: ~0.1ms (direct memory copy only)
-    /// - Different format/size with context reuse: ~2-5ms
-    /// - Different format/size with context reinit: ~5-10ms
-    ///
-    /// # Safety
-    ///
-    /// This function is unsafe because it directly manipulates FFmpeg AVFrame pointers.
-    /// The caller must ensure that `dst` is a valid, properly allocated AVFrame pointer
-    /// and that `codec_ctx` is a valid, open `AVCodecContext`.
-    unsafe fn convert_video_frame(
-        &mut self,
-        src: &VideoFrame,
-        dst: *mut AVFrame,
-        codec_ctx: *mut AVCodecContext,
-    ) -> Result<(), EncodeError> {
-        let target_fmt = (*codec_ctx).pix_fmt;
-        let target_width = (*codec_ctx).width as u32;
-        let target_height = (*codec_ctx).height as u32;
-
-        let src_fmt = pixel_format_to_av(src.format());
-        let src_width = src.width();
-        let src_height = src.height();
-
-        // Optimization 1: Skip conversion if format and dimensions match
-        if src_fmt == target_fmt && src_width == target_width && src_height == target_height {
-            return self.copy_frame_direct(src, dst, target_fmt);
-        }
-
-        // Optimization 2 & 3: Check if we need to reinitialize SwsContext
-        let needs_new_context = self.last_src_width != Some(src_width)
-            || self.last_src_height != Some(src_height)
-            || self.last_src_format != Some(src_fmt);
-
-        if needs_new_context || self.sws_ctx.is_none() {
-            // Free old context if exists
-            if let Some(ctx) = self.sws_ctx.take() {
-                swscale::free_context(ctx);
-            }
-
-            // Create new SwsContext with fast BILINEAR algorithm
-            let sws = swscale::get_context(
-                src_width as i32,
-                src_height as i32,
-                src_fmt,
-                target_width as i32,
-                target_height as i32,
-                target_fmt,
-                ff_sys::swscale::scale_flags::BILINEAR, // Fast scaling algorithm
-            )
-            .map_err(EncodeError::from_ffmpeg_error)?;
-
-            self.sws_ctx = Some(sws);
-            self.last_src_width = Some(src_width);
-            self.last_src_height = Some(src_height);
-            self.last_src_format = Some(src_fmt);
-        }
-
-        // Perform conversion using cached SwsContext
-        self.scale_frame(src, dst, target_fmt, target_width, target_height)
-    }
-
-    /// Copy frame data directly without scaling (when formats match).
-    unsafe fn copy_frame_direct(
-        &self,
-        src: &VideoFrame,
-        dst: *mut AVFrame,
-        target_fmt: AVPixelFormat,
-    ) -> Result<(), EncodeError> {
-        // Set frame properties
-        (*dst).format = target_fmt;
-        (*dst).width = src.width() as i32;
-        (*dst).height = src.height() as i32;
-
-        // Allocate frame buffer
-        let ret = ff_sys::av_frame_get_buffer(dst, 0);
-        if ret < 0 {
-            return Err(EncodeError::Ffmpeg {
-                code: ret,
-                message: format!(
-                    "Cannot allocate frame buffer: {}",
-                    ff_sys::av_error_string(ret)
-                ),
-            });
-        }
-
-        // Copy each plane directly
-        for (i, plane) in src.planes().iter().enumerate() {
-            if i >= (*dst).data.len() || (*dst).data[i].is_null() {
-                break;
-            }
-
-            // Bounds check for strides array
-            let src_stride = src
-                .strides()
-                .get(i)
-                .copied()
-                .ok_or_else(|| EncodeError::Ffmpeg {
-                    code: 0,
-                    message: format!("Missing stride for plane {}", i),
-                })?;
-
-            let dst_stride = (*dst).linesize[i] as usize;
-            let plane_data = plane.data();
-            let plane_height = self.get_plane_height(src.height(), i, src.format());
-
-            // Optimization: If strides match, copy entire plane at once
-            if src_stride == dst_stride {
-                let total_size = src_stride * plane_height;
-                if total_size <= plane_data.len() {
-                    ptr::copy_nonoverlapping(plane_data.as_ptr(), (*dst).data[i], total_size);
-                    continue;
-                }
-            }
-
-            // Copy line by line to handle different strides
-            for y in 0..plane_height {
-                let src_offset = y * src_stride;
-                let dst_offset = y * dst_stride;
-                let line_size = src_stride.min(dst_stride);
-
-                if src_offset + line_size <= plane_data.len() {
-                    ptr::copy_nonoverlapping(
-                        plane_data.as_ptr().add(src_offset),
-                        (*dst).data[i].add(dst_offset),
-                        line_size,
-                    );
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Scale frame using SwsContext (when formats or dimensions differ).
-    unsafe fn scale_frame(
-        &self,
-        src: &VideoFrame,
-        dst: *mut AVFrame,
-        target_fmt: AVPixelFormat,
-        target_width: u32,
-        target_height: u32,
-    ) -> Result<(), EncodeError> {
-        // Set frame properties
-        (*dst).format = target_fmt;
-        (*dst).width = target_width as i32;
-        (*dst).height = target_height as i32;
-
-        // Allocate frame buffer
-        let ret = ff_sys::av_frame_get_buffer(dst, 0);
-        if ret < 0 {
-            return Err(EncodeError::Ffmpeg {
-                code: ret,
-                message: format!(
-                    "Cannot allocate frame buffer: {}",
-                    ff_sys::av_error_string(ret)
-                ),
-            });
-        }
-
-        // Prepare source data pointers and strides
-        let mut src_data: [*const u8; MAX_PLANES] = [ptr::null(); MAX_PLANES];
-        let mut src_linesize: [i32; MAX_PLANES] = [0; MAX_PLANES];
-
-        for (i, plane) in src.planes().iter().enumerate() {
-            if i < MAX_PLANES {
-                src_data[i] = plane.data().as_ptr();
-                src_linesize[i] = src.strides()[i] as i32;
-            }
-        }
-
-        // Perform scaling/conversion
-        let sws_ctx = self.sws_ctx.ok_or_else(|| EncodeError::Ffmpeg {
-            code: 0,
-            message: "Scaling context not initialized".to_string(),
-        })?;
-
-        swscale::scale(
-            sws_ctx,
-            src_data.as_ptr(),
-            src_linesize.as_ptr(),
-            0,
-            src.height() as i32,
-            (*dst).data.as_mut_ptr().cast_const(),
-            (*dst).linesize.as_mut_ptr(),
-        )
-        .map_err(EncodeError::from_ffmpeg_error)?;
-
-        Ok(())
-    }
-
-    /// Calculate the height of a plane for a given frame height and pixel format.
-    ///
-    /// Different pixel formats have different plane heights. For YUV 4:2:0 formats,
-    /// the U/V planes are half the height of the Y plane.
-    ///
-    /// # Arguments
-    ///
-    /// * `frame_height` - The height of the entire frame
-    /// * `plane_index` - The plane index (0: Y/RGB, 1: U/UV, 2: V)
-    /// * `format` - The pixel format
-    ///
-    /// # Returns
-    ///
-    /// The height (number of rows) for the specified plane.
-    #[allow(clippy::manual_div_ceil)]
-    fn get_plane_height(
-        &self,
-        frame_height: u32,
-        plane_index: usize,
-        format: ff_format::PixelFormat,
-    ) -> usize {
-        use ff_format::PixelFormat;
-
-        match format {
-            // YUV 4:2:0 - U and V planes are half height
-            PixelFormat::Yuv420p | PixelFormat::Yuv420p10le => {
-                if plane_index == 0 {
-                    frame_height as usize
-                } else {
-                    // Safe division with ceiling: (height + 1) / 2
-                    // Equivalent to div_ceil(2) but more explicit about avoiding overflow
-                    // Note: div_ceil() internally uses (n + d - 1) / d which could overflow
-                    ((frame_height as usize) + 1) / 2
-                }
-            }
-            // Semi-planar NV12/NV21/P010 - UV plane is half height
-            PixelFormat::Nv12 | PixelFormat::Nv21 | PixelFormat::P010le => {
-                if plane_index == 0 {
-                    frame_height as usize
-                } else {
-                    // Safe division with ceiling: (height + 1) / 2
-                    ((frame_height as usize) + 1) / 2
-                }
-            }
-            // All other formats - full height for all planes
-            _ => frame_height as usize,
-        }
-    }
-
-    /// Receive encoded packets from the encoder.
-    unsafe fn receive_packets(&mut self) -> Result<(), EncodeError> {
-        let codec_ctx = self
-            .video_codec_ctx
-            .ok_or_else(|| EncodeError::InvalidConfig {
-                reason: "Video codec not initialized".to_string(),
-            })?;
-
-        let mut packet = av_packet_alloc();
-        if packet.is_null() {
-            return Err(EncodeError::Ffmpeg {
-                code: 0,
-                message: "Cannot allocate packet".to_string(),
-            });
-        }
-
-        loop {
-            match avcodec::receive_packet(codec_ctx, packet) {
-                Ok(()) => {
-                    // Packet received successfully
-                }
-                Err(e) if e == ff_sys::error_codes::EAGAIN || e == ff_sys::error_codes::EOF => {
-                    // No more packets available
-                    break;
-                }
-                Err(e) => {
-                    av_packet_free(&mut packet as *mut *mut _);
-                    return Err(EncodeError::Ffmpeg {
-                        code: e,
-                        message: format!("Error receiving packet: {}", ff_sys::av_error_string(e)),
-                    });
-                }
-            }
-
-            // Set stream index
-            (*packet).stream_index = self.video_stream_index;
-
-            // Attach HDR10 side data to keyframe packets.
-            if let Some(ref meta) = self.hdr10_metadata {
-                const AV_PKT_FLAG_KEY: i32 = 1;
-                if (*packet).flags & AV_PKT_FLAG_KEY != 0 {
-                    self.attach_hdr10_side_data(packet, meta);
-                }
-            }
-
-            // Write packet
-            let write_ret = av_interleaved_write_frame(self.format_ctx, packet);
-            if write_ret < 0 {
-                av_packet_unref(packet);
-                av_packet_free(&mut packet as *mut *mut _);
-                return Err(EncodeError::MuxingFailed {
-                    reason: ff_sys::av_error_string(write_ret),
-                });
-            }
-
-            self.bytes_written += (*packet).size as u64;
-
-            av_packet_unref(packet);
-        }
-
-        av_packet_free(&mut packet as *mut *mut _);
-        Ok(())
-    }
-
-    /// Attach HDR10 static metadata as packet side data to a keyframe packet.
-    ///
-    /// Attaches `AV_PKT_DATA_CONTENT_LIGHT_LEVEL` (MaxCLL/MaxFALL) and
-    /// `AV_PKT_DATA_MASTERING_DISPLAY_METADATA` to `pkt`.  Called from
-    /// `receive_packets` for every keyframe when `hdr10_metadata` is set.
-    ///
-    /// # Safety
-    ///
-    /// `pkt` must be a valid, non-null pointer to an allocated `AVPacket`.
-    unsafe fn attach_hdr10_side_data(&self, pkt: *mut AVPacket, meta: &ff_format::Hdr10Metadata) {
-        // ── Content light level (MaxCLL / MaxFALL) ──────────────────────────
-        let cll_ptr = av_packet_new_side_data(
-            pkt,
-            AVPacketSideDataType_AV_PKT_DATA_CONTENT_LIGHT_LEVEL,
-            std::mem::size_of::<AvContentLightMetadata>(),
-        );
-        if cll_ptr.is_null() {
-            log::warn!(
-                "hdr10 side_data allocation failed, skipping MaxCLL/MaxFALL type=content_light_level"
-            );
-        } else {
-            // SAFETY: av_packet_new_side_data returns memory allocated for exactly
-            // sizeof(AvContentLightMetadata) bytes with alignment suitable for the
-            // data type. We use ptr::write to avoid creating a reference to
-            // potentially-unaligned memory (ptr::write handles any alignment).
-            let cll_value = AvContentLightMetadata {
-                max_cll: u32::from(meta.max_cll),
-                max_fall: u32::from(meta.max_fall),
-            };
-            // SAFETY: write_unaligned does not require the pointer to be aligned;
-            // it copies bytes. FFmpeg's av_packet_new_side_data returns
-            // malloc-aligned memory in practice, but write_unaligned is correct
-            // regardless of alignment.
-            std::ptr::write_unaligned(cll_ptr.cast::<AvContentLightMetadata>(), cll_value);
-        }
-
-        // ── Mastering display colour volume ─────────────────────────────────
-        let md_ptr = av_packet_new_side_data(
-            pkt,
-            AVPacketSideDataType_AV_PKT_DATA_MASTERING_DISPLAY_METADATA,
-            std::mem::size_of::<AvMasteringDisplayMetadata>(),
-        );
-        if md_ptr.is_null() {
-            log::warn!(
-                "hdr10 side_data allocation failed, skipping mastering display type=mastering_display_metadata"
-            );
-        } else {
-            let d = &meta.mastering_display;
-            let r = |n: u16| ff_sys::AVRational {
-                num: i32::from(n),
-                den: 50000,
-            };
-            let lum = |n: u32| ff_sys::AVRational {
-                num: i32::try_from(n).unwrap_or(i32::MAX),
-                den: 10000,
-            };
-            // SAFETY: av_packet_new_side_data returns memory allocated for exactly
-            // sizeof(AvMasteringDisplayMetadata) bytes. We use ptr::write to safely
-            // write the struct without creating a potentially-unaligned reference.
-            let md_value = AvMasteringDisplayMetadata {
-                // Chromaticity coordinates: denominator = 50000 (CIE 1931 xy).
-                // Order: [R, G, B][x, y]
-                display_primaries: [
-                    [r(d.red_x), r(d.red_y)],
-                    [r(d.green_x), r(d.green_y)],
-                    [r(d.blue_x), r(d.blue_y)],
-                ],
-                white_point: [r(d.white_x), r(d.white_y)],
-                // Luminance: denominator = 10000.
-                min_luminance: lum(d.min_luminance),
-                max_luminance: lum(d.max_luminance),
-                has_primaries: 1,
-                has_luminance: 1,
-            };
-            // SAFETY: write_unaligned does not require the pointer to be aligned;
-            // it copies bytes. FFmpeg's av_packet_new_side_data returns
-            // malloc-aligned memory in practice, but write_unaligned is correct
-            // regardless of alignment.
-            std::ptr::write_unaligned(md_ptr.cast::<AvMasteringDisplayMetadata>(), md_value);
-        }
-    }
-
-    /// Push an audio frame for encoding.
-    pub(super) fn push_audio_frame(&mut self, frame: &AudioFrame) -> Result<(), EncodeError> {
-        // SAFETY: self is properly initialised; all raw FFmpeg pointers are valid and exclusively owned.
-        unsafe {
-            let codec_ctx = self
-                .audio_codec_ctx
-                .ok_or_else(|| EncodeError::InvalidConfig {
-                    reason: "Audio codec not initialized".to_string(),
-                })?;
-
-            // Allocate AVFrame
-            let mut av_frame = av_frame_alloc();
-            if av_frame.is_null() {
-                return Err(EncodeError::Ffmpeg {
-                    code: 0,
-                    message: "Cannot allocate frame".to_string(),
-                });
-            }
-
-            // Convert AudioFrame to AVFrame
-            let convert_result = self.convert_audio_frame(frame, av_frame);
-            if let Err(e) = convert_result {
-                av_frame_free(&mut av_frame as *mut *mut _);
-                return Err(e);
-            }
-
-            // Set frame properties
-            (*av_frame).pts = self.audio_sample_count as i64;
-
-            // Send frame to encoder
-            let send_result = avcodec::send_frame(codec_ctx, av_frame);
-            if let Err(e) = send_result {
-                av_frame_free(&mut av_frame as *mut *mut _);
-                return Err(EncodeError::Ffmpeg {
-                    code: e,
-                    message: format!("Failed to send audio frame: {}", ff_sys::av_error_string(e)),
-                });
-            }
-
-            // Receive packets
-            let receive_result = self.receive_audio_packets();
-
-            // Always cleanup the frame
-            av_frame_free(&mut av_frame as *mut *mut _);
-
-            // Check if receiving packets failed
-            receive_result?;
-
-            self.audio_sample_count += frame.samples() as u64;
-
-            Ok(())
-        } // unsafe
-    }
-
-    /// Convert AudioFrame to AVFrame with resampling if needed.
-    unsafe fn convert_audio_frame(
-        &mut self,
-        src: &AudioFrame,
-        dst: *mut AVFrame,
-    ) -> Result<(), EncodeError> {
-        let codec_ctx = self
-            .audio_codec_ctx
-            .ok_or_else(|| EncodeError::InvalidConfig {
-                reason: "Audio codec not initialized".to_string(),
-            })?;
-
-        let target_sample_rate = (*codec_ctx).sample_rate;
-        let target_format = (*codec_ctx).sample_fmt;
-        let target_ch_layout = &(*codec_ctx).ch_layout;
-
-        // Check if we need to resample
-        let src_sample_rate = src.sample_rate() as i32;
-        let src_format = sample_format_to_av(src.format());
-        let src_ch_layout = {
-            let mut layout = AVChannelLayout::default();
-            swresample::channel_layout::set_default(&mut layout, src.channels() as i32);
-            layout
-        };
-
-        let needs_resampling = src_sample_rate != target_sample_rate
-            || src_format != target_format
-            || !swresample::channel_layout::is_equal(&src_ch_layout, target_ch_layout);
-
-        if needs_resampling {
-            // Initialize resampler if needed
-            if self.swr_ctx.is_none() {
-                let swr_ctx = swresample::alloc_set_opts2(
-                    target_ch_layout,
-                    target_format,
-                    target_sample_rate,
-                    &src_ch_layout,
-                    src_format,
-                    src_sample_rate,
-                )
-                .map_err(EncodeError::from_ffmpeg_error)?;
-
-                swresample::init(swr_ctx).map_err(EncodeError::from_ffmpeg_error)?;
-                self.swr_ctx = Some(swr_ctx);
-            }
-
-            let swr_ctx = self.swr_ctx.ok_or_else(|| EncodeError::Ffmpeg {
-                code: 0,
-                message: "Resampling context not initialized".to_string(),
-            })?;
-
-            // Estimate output sample count
-            let out_samples = swresample::estimate_output_samples(
-                target_sample_rate,
-                src_sample_rate,
-                src.samples() as i32,
-            );
-
-            // Set frame properties
-            (*dst).format = target_format;
-            (*dst).sample_rate = target_sample_rate;
-            (*dst).nb_samples = out_samples;
-
-            // Copy target channel layout
-            swresample::channel_layout::copy(&mut (*dst).ch_layout, target_ch_layout)
-                .map_err(EncodeError::from_ffmpeg_error)?;
-
-            // Allocate frame buffer
-            let ret = ff_sys::av_frame_get_buffer(dst, 0);
-            if ret < 0 {
-                return Err(EncodeError::Ffmpeg {
-                    code: ret,
-                    message: format!(
-                        "Cannot allocate audio frame buffer: {}",
-                        ff_sys::av_error_string(ret)
-                    ),
-                });
-            }
-
-            // Prepare input pointers
-            let in_ptrs: Vec<*const u8> = if src.format().is_planar() {
-                // Planar: one pointer per channel
-                src.planes().iter().map(|p| p.as_ptr()).collect()
-            } else {
-                // Packed: single pointer
-                vec![src.planes()[0].as_ptr()]
-            };
-
-            // Convert
-            let samples_out = swresample::convert(
-                swr_ctx,
-                (*dst).data.as_mut_ptr().cast(),
-                out_samples,
-                in_ptrs.as_ptr(),
-                src.samples() as i32,
-            )
-            .map_err(EncodeError::from_ffmpeg_error)?;
-
-            (*dst).nb_samples = samples_out;
-        } else {
-            // No resampling needed, direct copy
-            (*dst).format = src_format;
-            (*dst).sample_rate = src_sample_rate;
-            (*dst).nb_samples = src.samples() as i32;
-
-            // Copy channel layout
-            swresample::channel_layout::copy(&mut (*dst).ch_layout, &src_ch_layout)
-                .map_err(EncodeError::from_ffmpeg_error)?;
-
-            // Allocate frame buffer
-            let ret = ff_sys::av_frame_get_buffer(dst, 0);
-            if ret < 0 {
-                return Err(EncodeError::Ffmpeg {
-                    code: ret,
-                    message: format!(
-                        "Cannot allocate audio frame buffer: {}",
-                        ff_sys::av_error_string(ret)
-                    ),
-                });
-            }
-
-            // Copy audio data
-            if src.format().is_planar() {
-                // Copy each plane
-                for (i, plane) in src.planes().iter().enumerate() {
-                    if i < (*dst).data.len() && !(*dst).data[i].is_null() {
-                        let size = plane.len();
-                        ptr::copy_nonoverlapping(plane.as_ptr(), (*dst).data[i], size);
-                    }
-                }
-            } else {
-                // Copy single packed buffer
-                if !(*dst).data[0].is_null() {
-                    let size = src.planes()[0].len();
-                    ptr::copy_nonoverlapping(src.planes()[0].as_ptr(), (*dst).data[0], size);
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Receive encoded audio packets from the encoder.
-    unsafe fn receive_audio_packets(&mut self) -> Result<(), EncodeError> {
-        let codec_ctx = self
-            .audio_codec_ctx
-            .ok_or_else(|| EncodeError::InvalidConfig {
-                reason: "Audio codec not initialized".to_string(),
-            })?;
-
-        let mut packet = av_packet_alloc();
-        if packet.is_null() {
-            return Err(EncodeError::Ffmpeg {
-                code: 0,
-                message: "Cannot allocate packet".to_string(),
-            });
-        }
-
-        loop {
-            match avcodec::receive_packet(codec_ctx, packet) {
-                Ok(()) => {
-                    // Packet received successfully
-                }
-                Err(e) if e == ff_sys::error_codes::EAGAIN || e == ff_sys::error_codes::EOF => {
-                    // No more packets available
-                    break;
-                }
-                Err(e) => {
-                    av_packet_free(&mut packet as *mut *mut _);
-                    return Err(EncodeError::Ffmpeg {
-                        code: e,
-                        message: format!(
-                            "Error receiving audio packet: {}",
-                            ff_sys::av_error_string(e)
-                        ),
-                    });
-                }
-            }
-
-            // Set stream index
-            (*packet).stream_index = self.audio_stream_index;
-
-            // Write packet
-            let write_ret = av_interleaved_write_frame(self.format_ctx, packet);
-            if write_ret < 0 {
-                av_packet_unref(packet);
-                av_packet_free(&mut packet as *mut *mut _);
-                return Err(EncodeError::MuxingFailed {
-                    reason: ff_sys::av_error_string(write_ret),
-                });
-            }
-
-            self.bytes_written += (*packet).size as u64;
-
-            av_packet_unref(packet);
-        }
-
-        av_packet_free(&mut packet as *mut *mut _);
-        Ok(())
-    }
-
-    /// Finish encoding and write trailer.
-    pub(super) fn finish(&mut self) -> Result<(), EncodeError> {
-        // SAFETY: self is properly initialised; all raw FFmpeg pointers are valid and exclusively owned.
-        unsafe {
-            // For two-pass, run the second pass now (handles flushing + trailer).
-            if self.two_pass {
-                return self.run_pass2();
-            }
-
-            // Single-pass: flush video encoder
-            if let Some(codec_ctx) = self.video_codec_ctx {
-                // Send NULL frame to flush
-                avcodec::send_frame(codec_ctx, ptr::null())
-                    .map_err(EncodeError::from_ffmpeg_error)?;
-                self.receive_packets()?;
-            }
-
-            // Flush audio encoder
-            if let Some(codec_ctx) = self.audio_codec_ctx {
-                // Send NULL frame to flush
-                avcodec::send_frame(codec_ctx, ptr::null())
-                    .map_err(EncodeError::from_ffmpeg_error)?;
-                self.receive_audio_packets()?;
-            }
-
-            // Write subtitle passthrough packets before trailer.
-            self.write_subtitle_packets()?;
-
-            // Write trailer
-            let ret = av_write_trailer(self.format_ctx);
-            if ret < 0 {
-                return Err(EncodeError::Ffmpeg {
-                    code: ret,
-                    message: format!("Cannot write trailer: {}", ff_sys::av_error_string(ret)),
-                });
-            }
-
-            Ok(())
-        } // unsafe
-    }
-
-    /// Run the second pass of two-pass encoding.
-    ///
-    /// 1. Flushes the pass-1 encoder and collects `stats_out`.
-    /// 2. Initialises a pass-2 codec context with `AV_CODEC_FLAG_PASS2` and
-    ///    the collected statistics.
-    /// 3. Opens the real output file and writes the container header.
-    /// 4. Re-encodes all buffered frames through the pass-2 context.
-    /// 5. Flushes the pass-2 encoder and writes the container trailer.
-    ///
-    /// # Safety
-    ///
-    /// Must only be called from `finish` when `self.two_pass` is `true`.
-    /// All FFmpeg resources must be valid at the point of the call.
-    unsafe fn run_pass2(&mut self) -> Result<(), EncodeError> {
-        // ── Step 1: Flush pass-1 encoder ────────────────────────────────────
-        let mut pass1_ctx = self
-            .pass1_codec_ctx
-            .ok_or_else(|| EncodeError::InvalidConfig {
-                reason: "Pass-1 codec context not available".to_string(),
-            })?;
-
-        // SAFETY: pass1_ctx is a valid open codec context.
-        if let Err(e) = avcodec::send_frame(pass1_ctx, ptr::null())
-            && e != ff_sys::error_codes::EOF
-        {
-            return Err(EncodeError::Ffmpeg {
-                code: e,
-                message: format!("pass1 flush send_frame: {}", ff_sys::av_error_string(e)),
-            });
-        }
-        self.drain_pass1_packets(pass1_ctx)?;
-
-        // ── Step 2: Collect stats_out ────────────────────────────────────────
-        // SAFETY: stats_out is either null or a valid C string owned by the
-        // codec context; it remains valid until avcodec_free_context is called.
-        let stats_str = if !(*pass1_ctx).stats_out.is_null() {
-            std::ffi::CStr::from_ptr((*pass1_ctx).stats_out)
-                .to_string_lossy()
-                .into_owned()
-        } else {
-            log::warn!(
-                "two-pass pass-1 produced no stats_out; pass-2 quality may not improve \
-                 codec={}",
-                self.actual_video_codec
-            );
-            String::new()
-        };
-        log::info!("two-pass pass-1 complete stats_len={}", stats_str.len());
-
-        // ── Step 3: Free pass-1 codec context ───────────────────────────────
-        // SAFETY: pass1_ctx is no longer needed; we own it exclusively.
-        avcodec::free_context(&mut pass1_ctx as *mut *mut _);
-        self.pass1_codec_ctx = None;
-
-        // ── Step 4: Set up pass-2 codec context ─────────────────────────────
-        let config = self
-            .two_pass_config
-            .take()
-            .ok_or_else(|| EncodeError::InvalidConfig {
-                reason: "Two-pass config not available for pass-2 initialisation".to_string(),
-            })?;
-
-        let output_path = config.path.clone();
-        self.init_pass2_codec_ctx(&config, &stats_str)?;
-
-        // ── Step 5: Open output file and write header ────────────────────────
-        match ff_sys::avformat::open_output(&output_path, ff_sys::avformat::avio_flags::WRITE) {
-            Ok(pb) => (*self.format_ctx).pb = pb,
-            Err(_) => {
-                return Err(EncodeError::CannotCreateFile { path: output_path });
-            }
-        }
-
-        Self::apply_movflags(self.format_ctx, config.container);
-        Self::apply_metadata(self.format_ctx, &config.metadata);
-        Self::apply_chapters(self.format_ctx, &config.chapters);
-        let ret = avformat_write_header(self.format_ctx, ptr::null_mut());
-        if ret < 0 {
-            return Err(EncodeError::Ffmpeg {
-                code: ret,
-                message: format!(
-                    "Cannot write header in pass 2: {}",
-                    ff_sys::av_error_string(ret)
-                ),
-            });
-        }
-
-        // ── Step 6: Re-encode all buffered frames ────────────────────────────
-        let frames = std::mem::take(&mut self.buffered_frames);
-        self.frame_count = 0;
-        for tf in &frames {
-            self.push_two_pass_frame(tf)?;
-        }
-
-        // ── Step 7: Flush pass-2 encoder and write trailer ───────────────────
-        if let Some(codec_ctx) = self.video_codec_ctx {
-            // SAFETY: codec_ctx is a valid open pass-2 codec context.
-            if let Err(e) = avcodec::send_frame(codec_ctx, ptr::null())
-                && e != ff_sys::error_codes::EOF
-            {
-                return Err(EncodeError::Ffmpeg {
-                    code: e,
-                    message: format!("pass2 flush send_frame: {}", ff_sys::av_error_string(e)),
-                });
-            }
-            self.receive_packets()?;
-        }
-
-        // Write subtitle passthrough packets before trailer.
-        self.write_subtitle_packets()?;
-
-        let ret = av_write_trailer(self.format_ctx);
-        if ret < 0 {
-            return Err(EncodeError::Ffmpeg {
-                code: ret,
-                message: format!("Cannot write trailer: {}", ff_sys::av_error_string(ret)),
-            });
-        }
-
-        Ok(())
-    }
-
-    /// Initialise the pass-2 video codec context.
-    ///
-    /// Mirrors the configuration performed in `init_video_encoder` but sets
-    /// `AV_CODEC_FLAG_PASS2` and assigns `stats_in` from the pass-1 statistics
-    /// string. Does **not** create a new AVStream — the stream was already
-    /// registered during `init_video_encoder` (pass 1).
-    ///
-    /// # Safety
-    ///
-    /// Must only be called from `run_pass2`. `self.format_ctx` must be valid.
-    unsafe fn init_pass2_codec_ctx(
-        &mut self,
-        config: &VideoEncoderConfig,
-        stats: &str,
-    ) -> Result<(), EncodeError> {
-        use crate::BitrateMode;
-        let width = config.video_width.unwrap_or(0);
-        let height = config.video_height.unwrap_or(0);
-        let fps = config.video_fps.unwrap_or(30.0);
-        let encoder_name = self.actual_video_codec.clone();
-
-        let c_encoder_name =
-            CString::new(encoder_name.as_str()).map_err(|_| EncodeError::Ffmpeg {
-                code: 0,
-                message: "Invalid encoder name for pass 2".to_string(),
-            })?;
-
-        let codec_ptr =
-            avcodec::find_encoder_by_name(c_encoder_name.as_ptr()).ok_or_else(|| {
-                EncodeError::NoSuitableEncoder {
-                    codec: encoder_name.clone(),
-                    tried: vec![encoder_name.clone()],
-                }
-            })?;
-
-        let codec_ctx =
-            avcodec::alloc_context3(codec_ptr).map_err(EncodeError::from_ffmpeg_error)?;
-
-        // Mirror the same codec configuration as pass 1.
-        (*codec_ctx).codec_id = codec_to_id(config.video_codec);
-        (*codec_ctx).width = width as i32;
-        (*codec_ctx).height = height as i32;
-        (*codec_ctx).time_base.num = 1;
-        (*codec_ctx).time_base.den = (fps * 1000.0) as i32;
-        (*codec_ctx).framerate.num = fps as i32;
-        (*codec_ctx).framerate.den = 1;
-        (*codec_ctx).pix_fmt = AVPixelFormat_AV_PIX_FMT_YUV420P;
-
-        match config.video_bitrate_mode.as_ref() {
-            Some(BitrateMode::Cbr(bps)) => {
-                (*codec_ctx).bit_rate = *bps as i64;
-            }
-            Some(BitrateMode::Vbr { target, max }) => {
-                (*codec_ctx).bit_rate = *target as i64;
-                (*codec_ctx).rc_max_rate = *max as i64;
-                (*codec_ctx).rc_buffer_size = (*max * 2) as i32;
-            }
-            Some(BitrateMode::Crf(q)) => {
-                let crf_str = CString::new(q.to_string()).map_err(|_| EncodeError::Ffmpeg {
-                    code: 0,
-                    message: "Invalid CRF value".to_string(),
-                })?;
-                // SAFETY: priv_data, option name, and value are all valid pointers.
-                let ret = ff_sys::av_opt_set(
-                    (*codec_ctx).priv_data,
-                    b"crf\0".as_ptr() as *const i8,
-                    crf_str.as_ptr(),
-                    0,
-                );
-                if ret < 0 {
-                    log::warn!(
-                        "crf option not supported by pass-2 encoder, falling back to default \
-                         encoder={encoder_name} crf={q}"
-                    );
-                    (*codec_ctx).bit_rate = 2_000_000;
-                }
-            }
-            None => {
-                (*codec_ctx).bit_rate = 2_000_000;
-            }
-        }
-
-        if encoder_name.contains("264") || encoder_name.contains("265") {
-            let preset_cstr =
-                CString::new(config.preset.as_str()).map_err(|_| EncodeError::Ffmpeg {
-                    code: 0,
-                    message: "Invalid preset value".to_string(),
-                })?;
-            // SAFETY: priv_data, option name, and value are all valid pointers.
-            let ret = ff_sys::av_opt_set(
-                (*codec_ctx).priv_data,
-                b"preset\0".as_ptr() as *const i8,
-                preset_cstr.as_ptr(),
-                0,
-            );
-            if ret < 0 {
-                log::warn!(
-                    "preset option not supported by pass-2 encoder, ignoring \
-                     encoder={encoder_name} preset={}",
-                    config.preset
-                );
-            }
-        }
-
-        // Apply per-codec options before opening the pass-2 codec context.
-        if let Some(opts) = config.codec_options.as_ref() {
-            // SAFETY: codec_ctx is valid and allocated; priv_data is set by
-            // avcodec_alloc_context3. Options are applied before avcodec_open2
-            // so they take effect during codec initialisation.
-            Self::apply_codec_options(codec_ctx, opts, &encoder_name);
-        }
-
-        // Apply explicit pixel format override for pass 2 (mirrors pass 1).
-        if let Some(fmt) = config.pixel_format.as_ref() {
-            // SAFETY: codec_ctx is valid and allocated; direct field write is safe.
-            (*codec_ctx).pix_fmt = pixel_format_to_av(*fmt);
-        }
-
-        // Apply HDR10 color context for pass 2 (mirrors pass 1).
-        if config.hdr10_metadata.is_some() {
-            // SAFETY: codec_ctx is valid and allocated; direct field writes are safe.
-            (*codec_ctx).color_primaries = ff_sys::AVColorPrimaries_AVCOL_PRI_BT2020;
-            (*codec_ctx).color_trc = ff_sys::AVColorTransferCharacteristic_AVCOL_TRC_SMPTEST2084;
-            (*codec_ctx).colorspace = ff_sys::AVColorSpace_AVCOL_SPC_BT2020_NCL;
-        }
-
-        // Apply explicit color overrides for pass 2 (mirrors pass 1; take priority over HDR10).
-        if let Some(cs) = config.color_space {
-            // SAFETY: codec_ctx is valid and allocated; direct field write is safe.
-            (*codec_ctx).colorspace = color_space_to_av(cs);
-        }
-        if let Some(trc) = config.color_transfer {
-            // SAFETY: codec_ctx is valid and allocated; direct field write is safe.
-            (*codec_ctx).color_trc = color_transfer_to_av(trc);
-        }
-        if let Some(cp) = config.color_primaries {
-            // SAFETY: codec_ctx is valid and allocated; direct field write is safe.
-            (*codec_ctx).color_primaries = color_primaries_to_av(cp);
-        }
-
-        // Set the pass-2 flag and provide stats_in.
-        // SAFETY: codec_ctx is a valid allocated (but not yet opened) context.
-        (*codec_ctx).flags |= AV_CODEC_FLAG_PASS2;
-
-        // Point stats_in to our owned CString (kept alive in self.stats_in_cstr
-        // until cleanup() nulls the pointer and drops it).
-        if !stats.is_empty() {
-            let stats_cstr = CString::new(stats).map_err(|_| EncodeError::Ffmpeg {
-                code: 0,
-                message: "Invalid stats string from pass 1".to_string(),
-            })?;
-            // SAFETY: stats_cstr.as_ptr() is valid for the lifetime of stats_cstr,
-            // which is stored in self.stats_in_cstr and dropped only after the codec
-            // context is freed in cleanup().
-            (*codec_ctx).stats_in = stats_cstr.as_ptr().cast_mut();
-            self.stats_in_cstr = Some(stats_cstr);
-        }
-
-        // Try to open the pass-2 codec with PASS2 flag. Some encoders (e.g. the
-        // native mpeg4 encoder without meaningful stats) do not support PASS2 and
-        // return AVERROR(EPERM). In that case, fall back to opening without the
-        // flag so the caller still gets a valid encoder and usable output.
-        if avcodec::open2(codec_ctx, codec_ptr, ptr::null_mut()).is_err() {
-            log::warn!(
-                "two-pass pass-2 codec rejected AV_CODEC_FLAG_PASS2, \
-                 falling back to single-pass mode codec={encoder_name}"
-            );
-            (*codec_ctx).flags &= !AV_CODEC_FLAG_PASS2;
-            (*codec_ctx).stats_in = ptr::null_mut();
-            self.stats_in_cstr = None;
-            avcodec::open2(codec_ctx, codec_ptr, ptr::null_mut()).map_err(|e| {
-                EncodeError::Ffmpeg {
-                    code: e,
-                    message: format!(
-                        "pass2 avcodec_open2 fallback: {}",
-                        ff_sys::av_error_string(e)
-                    ),
-                }
-            })?;
-        }
-        log::info!(
-            "two-pass pass-2 codec opened codec={encoder_name} width={width} height={height}"
-        );
-
-        self.video_codec_ctx = Some(codec_ctx);
-        Ok(())
-    }
-
-    /// Encode a single buffered YUV420P frame through the pass-2 codec context.
-    ///
-    /// The frame data was captured during pass 1 (already converted to YUV420P)
-    /// and is re-encoded here with the optimised pass-2 settings.
-    ///
-    /// # Safety
-    ///
-    /// Must only be called from `run_pass2`. `self.video_codec_ctx` and
-    /// `self.format_ctx` must be valid and the output file must be open.
-    unsafe fn push_two_pass_frame(&mut self, tf: &TwoPassFrame) -> Result<(), EncodeError> {
-        let codec_ctx = self
-            .video_codec_ctx
-            .ok_or_else(|| EncodeError::InvalidConfig {
-                reason: "Pass-2 codec context not initialized".to_string(),
-            })?;
-
-        let mut av_frame = av_frame_alloc();
-        if av_frame.is_null() {
-            return Err(EncodeError::Ffmpeg {
-                code: 0,
-                message: "Cannot allocate frame for pass 2".to_string(),
-            });
-        }
-
-        // Set frame format — always YUV420P (converted during pass 1).
-        (*av_frame).format = AVPixelFormat_AV_PIX_FMT_YUV420P;
-        (*av_frame).width = tf.width as i32;
-        (*av_frame).height = tf.height as i32;
-
-        // Allocate the frame buffer.
-        let ret = ff_sys::av_frame_get_buffer(av_frame, 0);
-        if ret < 0 {
-            av_frame_free(&mut av_frame as *mut *mut _);
-            return Err(EncodeError::Ffmpeg {
-                code: ret,
-                message: format!(
-                    "Cannot allocate pass-2 frame buffer: {}",
-                    ff_sys::av_error_string(ret)
-                ),
-            });
-        }
-
-        // Copy the buffered YUV420P data into the AVFrame.
-        let uv_height = (tf.height as usize).div_ceil(2);
-        for (plane_idx, (plane_data, src_stride)) in
-            tf.planes.iter().zip(tf.strides.iter()).enumerate()
-        {
-            if plane_idx >= 3 || (*av_frame).data[plane_idx].is_null() || plane_data.is_empty() {
-                break;
-            }
-            let dst_stride = (*av_frame).linesize[plane_idx] as usize;
-            let plane_height = if plane_idx == 0 {
-                tf.height as usize
-            } else {
-                uv_height
-            };
-
-            for row in 0..plane_height {
-                let src_off = row * src_stride;
-                let dst_off = row * dst_stride;
-                let copy_len = (*src_stride).min(dst_stride);
-
-                if src_off + copy_len <= plane_data.len() {
-                    // SAFETY: bounds checked above; both pointers are valid and
-                    // the regions do not overlap.
-                    ptr::copy_nonoverlapping(
-                        plane_data.as_ptr().add(src_off),
-                        (*av_frame).data[plane_idx].add(dst_off),
-                        copy_len,
-                    );
-                }
-            }
-        }
-
-        (*av_frame).pts = tf.pts;
-
-        // Send to pass-2 encoder.
-        let send_result = avcodec::send_frame(codec_ctx, av_frame);
-        if let Err(e) = send_result {
-            av_frame_free(&mut av_frame as *mut *mut _);
-            return Err(EncodeError::Ffmpeg {
-                code: e,
-                message: format!(
-                    "Failed to send frame to pass-2 encoder: {}",
-                    ff_sys::av_error_string(e)
-                ),
-            });
-        }
-
-        let receive_result = self.receive_packets();
-        av_frame_free(&mut av_frame as *mut *mut _);
-        receive_result?;
-
-        self.frame_count += 1;
-        Ok(())
-    }
-
-    /// Open the subtitle source file, find the requested stream, register an output subtitle
-    /// stream with copied codec parameters, and close the source.
-    ///
-    /// Stores `(source_path, source_stream_index, output_stream_index)` in
-    /// Register binary attachment streams in the output container.
-    ///
-    /// Each attachment is stored as an `AVMEDIA_TYPE_ATTACHMENT` stream with
-    /// `AV_CODEC_ID_BIN_DATA`. The attachment data is placed in `extradata`
-    /// so the MKV muxer can write it into the container's `Attachments` element.
-    ///
-    /// Failures per entry are non-fatal: a warning is logged and the entry is
-    /// skipped so the rest of encoding can continue.
-    ///
-    /// # Safety
-    ///
-    /// `self.format_ctx` must be a valid, non-null `AVFormatContext` pointer.
-    /// Must be called before `avformat_write_header`.
-    unsafe fn init_attachments(&mut self, attachments: &[(Vec<u8>, String, String)]) {
-        for (data, mime_type, filename) in attachments {
-            // Create a new stream for the attachment.
-            // SAFETY: format_ctx is valid; null codec means the muxer selects a default.
-            let out_stream = avformat_new_stream(self.format_ctx, std::ptr::null());
-            if out_stream.is_null() {
-                log::warn!("attachment: avformat_new_stream failed, skipping filename={filename}");
-                continue;
-            }
-
-            let codecpar = (*out_stream).codecpar;
-            (*codecpar).codec_type = ff_sys::AVMediaType_AVMEDIA_TYPE_ATTACHMENT;
-            (*codecpar).codec_id = ff_sys::AVCodecID_AV_CODEC_ID_BIN_DATA;
-
-            // Allocate extradata with FFmpeg's allocator so it can be freed by
-            // avcodec_parameters_free. The padding bytes are zeroed by av_mallocz.
-            let alloc_size = data.len() + ff_sys::AV_INPUT_BUFFER_PADDING_SIZE as usize;
-            let extradata = ff_sys::av_mallocz(alloc_size) as *mut u8;
-            if extradata.is_null() {
-                log::warn!(
-                    "attachment: av_mallocz failed for extradata, skipping filename={filename}"
-                );
-                continue;
-            }
-            // SAFETY: extradata has at least `data.len()` bytes; data slice is valid.
-            std::ptr::copy_nonoverlapping(data.as_ptr(), extradata, data.len());
-            (*codecpar).extradata = extradata;
-            (*codecpar).extradata_size = data.len() as i32;
-
-            // Set stream metadata so the muxer records the filename and MIME type.
-            let Ok(c_filename) = std::ffi::CString::new(filename.as_str()) else {
-                log::warn!("attachment: filename contains null byte, skipping filename={filename}");
-                continue;
-            };
-            let Ok(c_mime) = std::ffi::CString::new(mime_type.as_str()) else {
-                log::warn!(
-                    "attachment: mime_type contains null byte, skipping filename={filename}"
-                );
-                continue;
-            };
-            // SAFETY: out_stream->metadata pointer is valid (initialized by avformat_new_stream).
-            ff_sys::av_dict_set(
-                &mut (*out_stream).metadata,
-                b"filename\0".as_ptr() as *const i8,
-                c_filename.as_ptr(),
-                0,
-            );
-            ff_sys::av_dict_set(
-                &mut (*out_stream).metadata,
-                b"mimetype\0".as_ptr() as *const i8,
-                c_mime.as_ptr(),
-                0,
-            );
-
-            log::info!(
-                "attachment: registered filename={filename} mime={mime_type} size={}",
-                data.len()
-            );
-        }
-    }
-
-    /// `self.subtitle_passthrough` on success. On any failure it logs a warning and returns
-    /// without modifying state, so encoding can continue without subtitles.
-    ///
-    /// # Safety
-    ///
-    /// `self.format_ctx` must be a valid, non-null `AVFormatContext` pointer.
-    /// Must be called before `avformat_write_header`.
-    unsafe fn init_subtitle_passthrough(&mut self, source_path: &str, source_stream_index: usize) {
-        let path = std::path::Path::new(source_path);
-        let src_ctx = match ff_sys::avformat::open_input(path) {
-            Ok(ctx) => ctx,
-            Err(e) => {
-                log::warn!(
-                    "subtitle_passthrough: failed to open source file \
-                     path={source_path} error={}",
-                    ff_sys::av_error_string(e)
-                );
-                return;
-            }
-        };
-
-        if let Err(e) = ff_sys::avformat::find_stream_info(src_ctx) {
-            log::warn!(
-                "subtitle_passthrough: failed to find stream info \
-                 path={source_path} error={}",
-                ff_sys::av_error_string(e)
-            );
-            let mut src_ctx_ptr = src_ctx;
-            ff_sys::avformat::close_input(&mut src_ctx_ptr);
-            return;
-        }
-
-        let nb_streams = (*src_ctx).nb_streams as usize;
-        if source_stream_index >= nb_streams {
-            log::warn!(
-                "subtitle_passthrough: stream index out of range \
-                 index={source_stream_index} nb_streams={nb_streams}"
-            );
-            let mut src_ctx_ptr = src_ctx;
-            ff_sys::avformat::close_input(&mut src_ctx_ptr);
-            return;
-        }
-
-        // SAFETY: source_stream_index < nb_streams; streams is a valid array.
-        let in_stream = *(*src_ctx).streams.add(source_stream_index);
-
-        if (*(*in_stream).codecpar).codec_type != AVMediaType_AVMEDIA_TYPE_SUBTITLE {
-            log::warn!(
-                "subtitle_passthrough: stream at index {source_stream_index} \
-                 is not a subtitle stream"
-            );
-            let mut src_ctx_ptr = src_ctx;
-            ff_sys::avformat::close_input(&mut src_ctx_ptr);
-            return;
-        }
-
-        // Record the output stream index before adding the new stream.
-        let out_stream_index = (*self.format_ctx).nb_streams as i32;
-        // SAFETY: format_ctx is valid; null codec means the muxer selects a default.
-        let out_stream = avformat_new_stream(self.format_ctx, std::ptr::null());
-        if out_stream.is_null() {
-            log::warn!("subtitle_passthrough: avformat_new_stream failed");
-            let mut src_ctx_ptr = src_ctx;
-            ff_sys::avformat::close_input(&mut src_ctx_ptr);
-            return;
-        }
-
-        // SAFETY: out_stream and in_stream->codecpar are valid non-null pointers.
-        let ret = ff_sys::avcodec_parameters_copy((*out_stream).codecpar, (*in_stream).codecpar);
-        if ret < 0 {
-            log::warn!(
-                "subtitle_passthrough: avcodec_parameters_copy failed error={}",
-                ff_sys::av_error_string(ret)
-            );
-            let mut src_ctx_ptr = src_ctx;
-            ff_sys::avformat::close_input(&mut src_ctx_ptr);
-            return;
-        }
-
-        // Reset codec_tag so the muxer can pick the appropriate value for the container.
-        (*(*out_stream).codecpar).codec_tag = 0;
-
-        let mut src_ctx_ptr = src_ctx;
-        ff_sys::avformat::close_input(&mut src_ctx_ptr);
-
-        self.subtitle_passthrough = Some((
-            source_path.to_string(),
-            source_stream_index,
-            out_stream_index,
-        ));
-        log::info!(
-            "subtitle_passthrough: registered subtitle stream \
-             source={source_path} stream_index={source_stream_index} \
-             out_stream_index={out_stream_index}"
-        );
-    }
-
-    /// Re-open the subtitle source file, read all packets from the registered subtitle stream,
-    /// rescale their timestamps, and write them to the output.
-    ///
-    /// No-op if `self.subtitle_passthrough` is `None`.  On non-fatal errors (open failure,
-    /// read errors) it logs a warning and returns `Ok(())` so the caller can still write the
-    /// trailer.
-    ///
-    /// # Safety
-    ///
-    /// `self.format_ctx` must be valid. Must be called before `av_write_trailer`.
-    unsafe fn write_subtitle_packets(&mut self) -> Result<(), EncodeError> {
-        let Some((source_path, source_stream_index, out_stream_index)) =
-            self.subtitle_passthrough.clone()
-        else {
-            return Ok(());
-        };
-
-        let path = std::path::Path::new(&source_path);
-        let src_ctx = match ff_sys::avformat::open_input(path) {
-            Ok(ctx) => ctx,
-            Err(e) => {
-                log::warn!(
-                    "subtitle_passthrough: failed to re-open source file \
-                     path={source_path} error={}",
-                    ff_sys::av_error_string(e)
-                );
-                return Ok(());
-            }
-        };
-
-        if let Err(e) = ff_sys::avformat::find_stream_info(src_ctx) {
-            log::warn!(
-                "subtitle_passthrough: failed to find stream info on re-open \
-                 path={source_path} error={}",
-                ff_sys::av_error_string(e)
-            );
-            let mut src_ctx_ptr = src_ctx;
-            ff_sys::avformat::close_input(&mut src_ctx_ptr);
-            return Ok(());
-        }
-
-        // SAFETY: source_stream_index was validated in init_subtitle_passthrough.
-        let in_stream = *(*src_ctx).streams.add(source_stream_index);
-        let in_time_base = (*in_stream).time_base;
-
-        // SAFETY: out_stream_index was set by avformat_new_stream; format_ctx is valid.
-        let out_stream = *(*self.format_ctx).streams.add(out_stream_index as usize);
-        let out_time_base = (*out_stream).time_base;
-
-        let pkt = av_packet_alloc();
-        if pkt.is_null() {
-            let mut src_ctx_ptr = src_ctx;
-            ff_sys::avformat::close_input(&mut src_ctx_ptr);
-            return Err(EncodeError::Ffmpeg {
-                code: 0,
-                message: "subtitle_passthrough: av_packet_alloc failed".to_string(),
-            });
-        }
-
-        loop {
-            match ff_sys::avformat::read_frame(src_ctx, pkt) {
-                Err(e) if e == ff_sys::error_codes::EOF => break,
-                Err(e) => {
-                    log::warn!(
-                        "subtitle_passthrough: read_frame error, stopping \
-                         path={source_path} error={}",
-                        ff_sys::av_error_string(e)
-                    );
-                    break;
-                }
-                Ok(()) => {}
-            }
-
-            // Skip packets from other streams.
-            if (*pkt).stream_index != source_stream_index as i32 {
-                av_packet_unref(pkt);
-                continue;
-            }
-
-            // Rescale timestamps from the source stream's time base to the output stream's.
-            // SAFETY: pkt is valid; time bases are plain value types.
-            ff_sys::av_packet_rescale_ts(pkt, in_time_base, out_time_base);
-            (*pkt).stream_index = out_stream_index;
-
-            // SRT/subtitle packets typically carry only PTS (DTS is AV_NOPTS_VALUE).
-            // The matroska muxer requires a valid DTS for av_interleaved_write_frame;
-            // mirror PTS → DTS when DTS is absent so packets are not silently dropped.
-            if (*pkt).dts == i64::MIN {
-                (*pkt).dts = (*pkt).pts;
-            }
-
-            let write_ret = av_interleaved_write_frame(self.format_ctx, pkt);
-            if write_ret < 0 {
-                log::warn!(
-                    "subtitle_passthrough: av_interleaved_write_frame failed \
-                     error={}",
-                    ff_sys::av_error_string(write_ret)
-                );
-            }
-            av_packet_unref(pkt);
-        }
-
-        av_packet_free(&mut (pkt as *mut _) as *mut *mut _);
-        let mut src_ctx_ptr = src_ctx;
-        ff_sys::avformat::close_input(&mut src_ctx_ptr);
-
-        Ok(())
-    }
-
-    /// Cleanup FFmpeg resources.
-    unsafe fn cleanup(&mut self) {
-        // Free video codec context.
-        // For two-pass encoding, stats_in points into self.stats_in_cstr (Rust-owned).
-        // Null it out BEFORE avcodec_free_context so FFmpeg does not call av_free on it.
-        if let Some(mut ctx) = self.video_codec_ctx.take() {
-            (*ctx).stats_in = ptr::null_mut();
-            avcodec::free_context(&mut ctx as *mut *mut _);
-        }
-        // Drop the owned CString now that the codec context no longer references it.
-        self.stats_in_cstr = None;
-
-        // Free pass-1 codec context (only set in two-pass mode).
-        if let Some(mut ctx) = self.pass1_codec_ctx.take() {
-            avcodec::free_context(&mut ctx as *mut *mut _);
-        }
-
-        // Free audio codec context
-        if let Some(mut ctx) = self.audio_codec_ctx.take() {
-            avcodec::free_context(&mut ctx as *mut *mut _);
-        }
-
-        // Free scaling context
-        if let Some(ctx) = self.sws_ctx.take() {
-            swscale::free_context(ctx);
-        }
-
-        // Free resampling context
-        if let Some(mut ctx) = self.swr_ctx.take() {
-            swresample::free(&mut ctx as *mut *mut _);
-        }
-
-        // Close output file
-        if !self.format_ctx.is_null() {
-            if !(*self.format_ctx).pb.is_null() {
-                ff_sys::avformat::close_output(&mut (*self.format_ctx).pb);
-            }
-            avformat_free_context(self.format_ctx);
-            self.format_ctx = ptr::null_mut();
-        }
-    }
-}
-
-impl Drop for VideoEncoderInner {
-    fn drop(&mut self) {
-        // SAFETY: We own all the FFmpeg resources and need to free them
-        unsafe {
-            self.cleanup();
-        }
-    }
-}
-
-// Helper functions
-
-fn codec_to_id(codec: VideoCodec) -> AVCodecID {
-    match codec {
-        VideoCodec::H264 => AVCodecID_AV_CODEC_ID_H264,
-        VideoCodec::H265 => AVCodecID_AV_CODEC_ID_HEVC,
-        VideoCodec::Vp9 => AVCodecID_AV_CODEC_ID_VP9,
-        VideoCodec::Av1 => AVCodecID_AV_CODEC_ID_AV1,
-        VideoCodec::Av1Svt => AVCodecID_AV_CODEC_ID_AV1,
-        VideoCodec::ProRes => AVCodecID_AV_CODEC_ID_PRORES,
-        VideoCodec::DnxHd => AVCodecID_AV_CODEC_ID_DNXHD,
-        VideoCodec::Mpeg4 => AVCodecID_AV_CODEC_ID_MPEG4,
-        VideoCodec::Vp8 => AVCodecID_AV_CODEC_ID_VP8,
-        VideoCodec::Mpeg2 => AVCodecID_AV_CODEC_ID_MPEG2VIDEO,
-        VideoCodec::Mjpeg => AVCodecID_AV_CODEC_ID_MJPEG,
-        VideoCodec::Png => AVCodecID_AV_CODEC_ID_PNG,
-        _ => AVCodecID_AV_CODEC_ID_NONE,
-    }
-}
-
-pub(super) fn preset_to_string(preset: &crate::Preset) -> String {
-    match preset {
-        crate::Preset::Ultrafast => "ultrafast",
-        crate::Preset::Faster => "faster",
-        crate::Preset::Fast => "fast",
-        crate::Preset::Medium => "medium",
-        crate::Preset::Slow => "slow",
-        crate::Preset::Slower => "slower",
-        crate::Preset::Veryslow => "veryslow",
-    }
-    .to_string()
-}
-
-/// Convert AudioCodec to FFmpeg AVCodecID.
-fn audio_codec_to_id(codec: AudioCodec) -> AVCodecID {
-    match codec {
-        AudioCodec::Aac => AVCodecID_AV_CODEC_ID_AAC,
-        AudioCodec::Opus => AVCodecID_AV_CODEC_ID_OPUS,
-        AudioCodec::Mp3 => AVCodecID_AV_CODEC_ID_MP3,
-        AudioCodec::Flac => AVCodecID_AV_CODEC_ID_FLAC,
-        AudioCodec::Pcm => AVCodecID_AV_CODEC_ID_PCM_S16LE,
-        AudioCodec::Pcm16 => AVCodecID_AV_CODEC_ID_PCM_S16LE,
-        AudioCodec::Pcm24 => AVCodecID_AV_CODEC_ID_PCM_S24LE,
-        AudioCodec::Vorbis => AVCodecID_AV_CODEC_ID_VORBIS,
-        AudioCodec::Ac3 => AVCodecID_AV_CODEC_ID_AC3,
-        AudioCodec::Eac3 => AVCodecID_AV_CODEC_ID_EAC3,
-        AudioCodec::Dts => AVCodecID_AV_CODEC_ID_DTS,
-        AudioCodec::Alac => AVCodecID_AV_CODEC_ID_ALAC,
-        _ => AVCodecID_AV_CODEC_ID_NONE,
-    }
-}
-
-/// Convert ff-format SampleFormat to FFmpeg AVSampleFormat.
-fn sample_format_to_av(format: ff_format::SampleFormat) -> ff_sys::AVSampleFormat {
-    use ff_format::SampleFormat;
-    use ff_sys::swresample::sample_format;
-
-    match format {
-        SampleFormat::U8 => sample_format::U8,
-        SampleFormat::I16 => sample_format::S16,
-        SampleFormat::I32 => sample_format::S32,
-        SampleFormat::F32 => sample_format::FLT,
-        SampleFormat::F64 => sample_format::DBL,
-        SampleFormat::U8p => sample_format::U8P,
-        SampleFormat::I16p => sample_format::S16P,
-        SampleFormat::I32p => sample_format::S32P,
-        SampleFormat::F32p => sample_format::FLTP,
-        SampleFormat::F64p => sample_format::DBLP,
-        _ => {
-            log::warn!(
-                "sample_format has no AV mapping, falling back to FLTP \
-                 format={format:?} fallback=FLTP"
-            );
-            sample_format::FLTP
-        }
-    }
-}
-
-/// Convert ff-format PixelFormat to FFmpeg AVPixelFormat.
-fn pixel_format_to_av(format: ff_format::PixelFormat) -> AVPixelFormat {
-    use ff_format::PixelFormat;
-
-    match format {
-        PixelFormat::Yuv420p => ff_sys::AVPixelFormat_AV_PIX_FMT_YUV420P,
-        PixelFormat::Yuv422p => ff_sys::AVPixelFormat_AV_PIX_FMT_YUV422P,
-        PixelFormat::Yuv444p => ff_sys::AVPixelFormat_AV_PIX_FMT_YUV444P,
-        PixelFormat::Rgb24 => ff_sys::AVPixelFormat_AV_PIX_FMT_RGB24,
-        PixelFormat::Bgr24 => ff_sys::AVPixelFormat_AV_PIX_FMT_BGR24,
-        PixelFormat::Rgba => ff_sys::AVPixelFormat_AV_PIX_FMT_RGBA,
-        PixelFormat::Bgra => ff_sys::AVPixelFormat_AV_PIX_FMT_BGRA,
-        PixelFormat::Gray8 => ff_sys::AVPixelFormat_AV_PIX_FMT_GRAY8,
-        PixelFormat::Nv12 => ff_sys::AVPixelFormat_AV_PIX_FMT_NV12,
-        PixelFormat::Nv21 => ff_sys::AVPixelFormat_AV_PIX_FMT_NV21,
-        PixelFormat::Yuv420p10le => ff_sys::AVPixelFormat_AV_PIX_FMT_YUV420P10LE,
-        PixelFormat::Yuv422p10le => ff_sys::AVPixelFormat_AV_PIX_FMT_YUV422P10LE,
-        PixelFormat::Yuv444p10le => ff_sys::AVPixelFormat_AV_PIX_FMT_YUV444P10LE,
-        PixelFormat::Yuva444p10le => ff_sys::AVPixelFormat_AV_PIX_FMT_YUVA444P10LE,
-        PixelFormat::P010le => ff_sys::AVPixelFormat_AV_PIX_FMT_P010LE,
-        _ => {
-            log::warn!(
-                "pixel_format has no AV mapping, falling back to Yuv420p \
-                 format={format:?} fallback=AV_PIX_FMT_YUV420P"
-            );
-            ff_sys::AVPixelFormat_AV_PIX_FMT_YUV420P
-        }
-    }
-}
-
-/// Convert FFmpeg AVPixelFormat back to ff-format PixelFormat.
-fn from_av_pixel_format(fmt: AVPixelFormat) -> ff_format::PixelFormat {
-    use ff_format::PixelFormat;
-    if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_YUV420P {
-        PixelFormat::Yuv420p
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_YUV422P {
-        PixelFormat::Yuv422p
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_YUV444P {
-        PixelFormat::Yuv444p
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_RGB24 {
-        PixelFormat::Rgb24
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_BGR24 {
-        PixelFormat::Bgr24
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_RGBA {
-        PixelFormat::Rgba
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_BGRA {
-        PixelFormat::Bgra
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_GRAY8 {
-        PixelFormat::Gray8
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_NV12 {
-        PixelFormat::Nv12
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_NV21 {
-        PixelFormat::Nv21
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_YUV420P10LE {
-        PixelFormat::Yuv420p10le
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_YUV422P10LE {
-        PixelFormat::Yuv422p10le
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_YUV444P10LE {
-        PixelFormat::Yuv444p10le
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_YUVA444P10LE {
-        PixelFormat::Yuva444p10le
-    } else if fmt == ff_sys::AVPixelFormat_AV_PIX_FMT_P010LE {
-        PixelFormat::P010le
-    } else {
-        log::warn!(
-            "pixel_format unsupported, falling back to Yuv420p requested={fmt} fallback=Yuv420p"
-        );
-        PixelFormat::Yuv420p
-    }
-}
-
-/// Convert ff-format ColorSpace to the FFmpeg AVColorSpace constant.
-fn color_space_to_av(cs: ff_format::ColorSpace) -> ff_sys::AVColorSpace {
-    use ff_format::ColorSpace;
-    match cs {
-        ColorSpace::Bt709 => ff_sys::AVColorSpace_AVCOL_SPC_BT709,
-        ColorSpace::Bt601 => ff_sys::AVColorSpace_AVCOL_SPC_SMPTE170M,
-        ColorSpace::Bt2020 => ff_sys::AVColorSpace_AVCOL_SPC_BT2020_NCL,
-        ColorSpace::DciP3 | ColorSpace::Srgb => ff_sys::AVColorSpace_AVCOL_SPC_RGB,
-        ColorSpace::Unknown => ff_sys::AVColorSpace_AVCOL_SPC_UNSPECIFIED,
-        _ => ff_sys::AVColorSpace_AVCOL_SPC_UNSPECIFIED,
-    }
-}
-
-/// Convert ff-format ColorTransfer to the FFmpeg AVColorTransferCharacteristic constant.
-fn color_transfer_to_av(trc: ff_format::ColorTransfer) -> ff_sys::AVColorTransferCharacteristic {
-    use ff_format::ColorTransfer;
-    match trc {
-        ColorTransfer::Bt709 => ff_sys::AVColorTransferCharacteristic_AVCOL_TRC_BT709,
-        ColorTransfer::Bt2020_10 => ff_sys::AVColorTransferCharacteristic_AVCOL_TRC_BT2020_10,
-        ColorTransfer::Bt2020_12 => ff_sys::AVColorTransferCharacteristic_AVCOL_TRC_BT2020_12,
-        ColorTransfer::Hlg => ff_sys::AVColorTransferCharacteristic_AVCOL_TRC_ARIB_STD_B67,
-        ColorTransfer::Pq => ff_sys::AVColorTransferCharacteristic_AVCOL_TRC_SMPTEST2084,
-        ColorTransfer::Linear => ff_sys::AVColorTransferCharacteristic_AVCOL_TRC_LINEAR,
-        ColorTransfer::Unknown => ff_sys::AVColorTransferCharacteristic_AVCOL_TRC_UNSPECIFIED,
-        _ => ff_sys::AVColorTransferCharacteristic_AVCOL_TRC_UNSPECIFIED,
-    }
-}
-
-/// Convert ff-format ColorPrimaries to the FFmpeg AVColorPrimaries constant.
-fn color_primaries_to_av(cp: ff_format::ColorPrimaries) -> ff_sys::AVColorPrimaries {
-    use ff_format::ColorPrimaries;
-    match cp {
-        ColorPrimaries::Bt709 => ff_sys::AVColorPrimaries_AVCOL_PRI_BT709,
-        ColorPrimaries::Bt601 => ff_sys::AVColorPrimaries_AVCOL_PRI_SMPTE170M,
-        ColorPrimaries::Bt2020 => ff_sys::AVColorPrimaries_AVCOL_PRI_BT2020,
-        ColorPrimaries::Unknown => ff_sys::AVColorPrimaries_AVCOL_PRI_UNSPECIFIED,
-        _ => ff_sys::AVColorPrimaries_AVCOL_PRI_UNSPECIFIED,
-    }
-}
-
-// SAFETY: VideoEncoderInner owns all FFmpeg contexts exclusively.
-//         These contexts are not accessed from multiple threads simultaneously;
-//         all access is serialized by whichever thread holds the VideoEncoder.
-//         Ownership transfer between threads is safe because FFmpeg contexts
-//         are created and destroyed on the same thread (via std::thread::spawn).
-unsafe impl Send for VideoEncoderInner {}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_h264_encoder_candidates_auto() {
-        let inner = create_dummy_encoder_inner();
-        let candidates = inner.select_h264_encoder_candidates(crate::HardwareEncoder::Auto);
-
-        // Should include hardware encoders
-        #[cfg(feature = "hwaccel")]
-        {
-            assert!(candidates.contains(&"h264_nvenc"));
-            assert!(candidates.contains(&"h264_qsv"));
-        }
-
-        // Should include libx264 if GPL feature is enabled
-        #[cfg(feature = "gpl")]
-        {
-            assert!(candidates.contains(&"libx264"));
-        }
-
-        // Should always include VP9 fallback
-        assert!(candidates.contains(&"libvpx-vp9"));
-    }
-
-    #[test]
-    fn test_h264_encoder_candidates_nvenc() {
-        let inner = create_dummy_encoder_inner();
-        let candidates = inner.select_h264_encoder_candidates(crate::HardwareEncoder::Nvenc);
-
-        #[cfg(feature = "hwaccel")]
-        {
-            // NVENC should be first priority
-            assert_eq!(candidates[0], "h264_nvenc");
-        }
-
-        // Should include VP9 fallback
-        assert!(candidates.contains(&"libvpx-vp9"));
-    }
-
-    #[test]
-    fn test_h264_encoder_candidates_none() {
-        let inner = create_dummy_encoder_inner();
-        let candidates = inner.select_h264_encoder_candidates(crate::HardwareEncoder::None);
-
-        #[cfg(feature = "hwaccel")]
-        {
-            // Should not include hardware encoders
-            assert!(!candidates.contains(&"h264_nvenc"));
-            assert!(!candidates.contains(&"h264_qsv"));
-        }
-
-        // Should include VP9 fallback
-        assert!(candidates.contains(&"libvpx-vp9"));
-    }
-
-    #[test]
-    fn test_h265_encoder_candidates_auto() {
-        let inner = create_dummy_encoder_inner();
-        let candidates = inner.select_h265_encoder_candidates(crate::HardwareEncoder::Auto);
-
-        // Should include hardware encoders
-        #[cfg(feature = "hwaccel")]
-        {
-            assert!(candidates.contains(&"hevc_nvenc"));
-            assert!(candidates.contains(&"hevc_qsv"));
-        }
-
-        // Should include libx265 if GPL feature is enabled
-        #[cfg(feature = "gpl")]
-        {
-            assert!(candidates.contains(&"libx265"));
-        }
-
-        // Should always include AV1 fallback
-        assert!(candidates.contains(&"libaom-av1") || candidates.contains(&"libsvtav1"));
-    }
-
-    #[test]
-    fn test_lgpl_fallback_priority() {
-        let inner = create_dummy_encoder_inner();
-
-        // Test H264 candidates
-        let h264_candidates = inner.select_h264_encoder_candidates(crate::HardwareEncoder::None);
-
-        #[cfg(not(feature = "gpl"))]
-        {
-            // Without GPL feature, should only have VP9
-            assert_eq!(h264_candidates, vec!["libvpx-vp9"]);
-        }
-
-        // Test H265 candidates
-        let h265_candidates = inner.select_h265_encoder_candidates(crate::HardwareEncoder::None);
-
-        #[cfg(not(feature = "gpl"))]
-        {
-            // Without GPL feature, should only have AV1 options
-            assert!(h265_candidates.contains(&"libaom-av1"));
-            assert!(!h265_candidates.contains(&"libx265"));
-        }
-    }
-
-    #[test]
-    fn test_get_plane_height_yuv420p() {
-        let inner = create_dummy_encoder_inner();
-
-        // Test YUV420P format - Y plane is full height, U/V planes are half height
-        // Even height (640x480)
-        assert_eq!(
-            inner.get_plane_height(480, 0, ff_format::PixelFormat::Yuv420p),
-            480
-        );
-        assert_eq!(
-            inner.get_plane_height(480, 1, ff_format::PixelFormat::Yuv420p),
-            240
-        );
-        assert_eq!(
-            inner.get_plane_height(480, 2, ff_format::PixelFormat::Yuv420p),
-            240
-        );
-
-        // Odd height (641x481) - test ceiling division
-        assert_eq!(
-            inner.get_plane_height(481, 0, ff_format::PixelFormat::Yuv420p),
-            481
-        );
-        assert_eq!(
-            inner.get_plane_height(481, 1, ff_format::PixelFormat::Yuv420p),
-            241
-        ); // (481 + 1) / 2 = 241
-        assert_eq!(
-            inner.get_plane_height(481, 2, ff_format::PixelFormat::Yuv420p),
-            241
-        );
-    }
-
-    #[test]
-    fn test_get_plane_height_nv12() {
-        let inner = create_dummy_encoder_inner();
-
-        // Test NV12 format - Y plane is full height, UV plane is half height
-        assert_eq!(
-            inner.get_plane_height(1080, 0, ff_format::PixelFormat::Nv12),
-            1080
-        );
-        assert_eq!(
-            inner.get_plane_height(1080, 1, ff_format::PixelFormat::Nv12),
-            540
-        );
-
-        // Odd height
-        assert_eq!(
-            inner.get_plane_height(1081, 0, ff_format::PixelFormat::Nv12),
-            1081
-        );
-        assert_eq!(
-            inner.get_plane_height(1081, 1, ff_format::PixelFormat::Nv12),
-            541
-        ); // (1081 + 1) / 2 = 541
-    }
-
-    #[test]
-    fn test_get_plane_height_rgba() {
-        let inner = create_dummy_encoder_inner();
-
-        // Test RGBA format - all planes are full height (only 1 plane)
-        assert_eq!(
-            inner.get_plane_height(720, 0, ff_format::PixelFormat::Rgba),
-            720
-        );
-        assert_eq!(
-            inner.get_plane_height(720, 1, ff_format::PixelFormat::Rgba),
-            720
-        );
-    }
-
-    #[test]
-    fn test_pixel_format_to_av() {
-        // Test common pixel formats
-        assert_eq!(
-            pixel_format_to_av(ff_format::PixelFormat::Yuv420p),
-            ff_sys::AVPixelFormat_AV_PIX_FMT_YUV420P
-        );
-        assert_eq!(
-            pixel_format_to_av(ff_format::PixelFormat::Rgba),
-            ff_sys::AVPixelFormat_AV_PIX_FMT_RGBA
-        );
-        assert_eq!(
-            pixel_format_to_av(ff_format::PixelFormat::Nv12),
-            ff_sys::AVPixelFormat_AV_PIX_FMT_NV12
-        );
-
-        // Test fallback for Other format
-        assert_eq!(
-            pixel_format_to_av(ff_format::PixelFormat::Other(999)),
-            ff_sys::AVPixelFormat_AV_PIX_FMT_YUV420P
-        );
-
-        // 10-bit formats
-        assert_eq!(
-            pixel_format_to_av(ff_format::PixelFormat::Yuv420p10le),
-            ff_sys::AVPixelFormat_AV_PIX_FMT_YUV420P10LE
-        );
-        assert_eq!(
-            pixel_format_to_av(ff_format::PixelFormat::Yuv422p10le),
-            ff_sys::AVPixelFormat_AV_PIX_FMT_YUV422P10LE
-        );
-        assert_eq!(
-            pixel_format_to_av(ff_format::PixelFormat::Yuv444p10le),
-            ff_sys::AVPixelFormat_AV_PIX_FMT_YUV444P10LE
-        );
-    }
-
-    #[test]
-    fn from_av_pixel_format_yuv420p10le_should_return_yuv420p10le() {
-        assert_eq!(
-            from_av_pixel_format(ff_sys::AVPixelFormat_AV_PIX_FMT_YUV420P10LE),
-            ff_format::PixelFormat::Yuv420p10le
-        );
-    }
-
-    #[test]
-    fn from_av_pixel_format_yuv422p10le_should_return_yuv422p10le() {
-        assert_eq!(
-            from_av_pixel_format(ff_sys::AVPixelFormat_AV_PIX_FMT_YUV422P10LE),
-            ff_format::PixelFormat::Yuv422p10le
-        );
-    }
-
-    #[test]
-    fn from_av_pixel_format_yuv444p10le_should_return_yuv444p10le() {
-        assert_eq!(
-            from_av_pixel_format(ff_sys::AVPixelFormat_AV_PIX_FMT_YUV444P10LE),
-            ff_format::PixelFormat::Yuv444p10le
-        );
-    }
-
-    #[test]
-    fn from_av_pixel_format_unknown_should_fall_back_to_yuv420p() {
-        assert_eq!(
-            from_av_pixel_format(ff_sys::AVPixelFormat_AV_PIX_FMT_NONE),
-            ff_format::PixelFormat::Yuv420p
-        );
-    }
-
-    #[test]
-    fn color_space_to_av_bt709_should_return_bt709() {
-        assert_eq!(
-            color_space_to_av(ff_format::ColorSpace::Bt709),
-            ff_sys::AVColorSpace_AVCOL_SPC_BT709
-        );
-    }
-
-    #[test]
-    fn color_space_to_av_bt2020_should_return_bt2020_ncl() {
-        assert_eq!(
-            color_space_to_av(ff_format::ColorSpace::Bt2020),
-            ff_sys::AVColorSpace_AVCOL_SPC_BT2020_NCL
-        );
-    }
-
-    #[test]
-    fn color_space_to_av_dcip3_should_return_rgb() {
-        assert_eq!(
-            color_space_to_av(ff_format::ColorSpace::DciP3),
-            ff_sys::AVColorSpace_AVCOL_SPC_RGB
-        );
-    }
-
-    #[test]
-    fn color_transfer_to_av_hlg_should_return_arib_std_b67() {
-        assert_eq!(
-            color_transfer_to_av(ff_format::ColorTransfer::Hlg),
-            ff_sys::AVColorTransferCharacteristic_AVCOL_TRC_ARIB_STD_B67
-        );
-    }
-
-    #[test]
-    fn color_transfer_to_av_pq_should_return_smptest2084() {
-        assert_eq!(
-            color_transfer_to_av(ff_format::ColorTransfer::Pq),
-            ff_sys::AVColorTransferCharacteristic_AVCOL_TRC_SMPTEST2084
-        );
-    }
-
-    #[test]
-    fn color_transfer_to_av_bt709_should_return_bt709() {
-        assert_eq!(
-            color_transfer_to_av(ff_format::ColorTransfer::Bt709),
-            ff_sys::AVColorTransferCharacteristic_AVCOL_TRC_BT709
-        );
-    }
-
-    #[test]
-    fn color_primaries_to_av_bt2020_should_return_bt2020() {
-        assert_eq!(
-            color_primaries_to_av(ff_format::ColorPrimaries::Bt2020),
-            ff_sys::AVColorPrimaries_AVCOL_PRI_BT2020
-        );
-    }
-
-    #[test]
-    fn color_primaries_to_av_bt709_should_return_bt709() {
-        assert_eq!(
-            color_primaries_to_av(ff_format::ColorPrimaries::Bt709),
-            ff_sys::AVColorPrimaries_AVCOL_PRI_BT709
-        );
-    }
-
-    #[test]
-    fn test_sws_context_tracking() {
-        let mut inner = create_dummy_encoder_inner();
-
-        // Initially no context
-        assert_eq!(inner.last_src_width, None);
-        assert_eq!(inner.last_src_height, None);
-        assert_eq!(inner.last_src_format, None);
-
-        // After setting (simulating what convert_video_frame does)
-        inner.last_src_width = Some(1920);
-        inner.last_src_height = Some(1080);
-        inner.last_src_format = Some(ff_sys::AVPixelFormat_AV_PIX_FMT_YUV420P);
-
-        // Verify tracking
-        assert_eq!(inner.last_src_width, Some(1920));
-        assert_eq!(inner.last_src_height, Some(1080));
-        assert_eq!(
-            inner.last_src_format,
-            Some(ff_sys::AVPixelFormat_AV_PIX_FMT_YUV420P)
-        );
-    }
-
-    /// Helper function to create a dummy encoder inner for testing.
-    fn create_dummy_encoder_inner() -> VideoEncoderInner {
-        VideoEncoderInner {
-            format_ctx: ptr::null_mut(),
-            video_codec_ctx: None,
-            audio_codec_ctx: None,
-            video_stream_index: -1,
-            audio_stream_index: -1,
-            sws_ctx: None,
-            swr_ctx: None,
-            frame_count: 0,
-            audio_sample_count: 0,
-            bytes_written: 0,
-            actual_video_codec: String::new(),
-            actual_audio_codec: String::new(),
-            last_src_width: None,
-            last_src_height: None,
-            last_src_format: None,
-            two_pass: false,
-            pass1_codec_ctx: None,
-            buffered_frames: Vec::new(),
-            two_pass_config: None,
-            stats_in_cstr: None,
-            subtitle_passthrough: None,
-            hdr10_metadata: None,
-        }
-    }
-}
